Enum binjs_es6::ast::ViewMutExpressionOrTemplateElement

pub enum ViewMutExpressionOrTemplateElement<'a> {
    LiteralBooleanExpression(&'a mut LiteralBooleanExpression),
    ObjectExpression(&'a mut ObjectExpression),
    ComputedMemberExpression(&'a mut ComputedMemberExpression),
    NewTargetExpression(&'a mut NewTargetExpression),
    AwaitExpression(&'a mut AwaitExpression),
    StaticMemberExpression(&'a mut StaticMemberExpression),
    LazyArrowExpressionWithExpression(&'a mut LazyArrowExpressionWithExpression),
    ThisExpression(&'a mut ThisExpression),
    EagerArrowExpressionWithExpression(&'a mut EagerArrowExpressionWithExpression),
    LazyFunctionExpression(&'a mut LazyFunctionExpression),
    LiteralRegExpExpression(&'a mut LiteralRegExpExpression),
    ConditionalExpression(&'a mut ConditionalExpression),
    BinaryExpression(&'a mut BinaryExpression),
    BinASTExpressionWithProbabilityTable(&'a mut BinASTExpressionWithProbabilityTable),
    YieldExpression(&'a mut YieldExpression),
    ArrayExpression(&'a mut ArrayExpression),
    ClassExpression(&'a mut ClassExpression),
    TemplateExpression(&'a mut TemplateExpression),
    EagerFunctionExpression(&'a mut EagerFunctionExpression),
    LiteralInfinityExpression(&'a mut LiteralInfinityExpression),
    LiteralStringExpression(&'a mut LiteralStringExpression),
    CompoundAssignmentExpression(&'a mut CompoundAssignmentExpression),
    TemplateElement(&'a mut TemplateElement),
    LiteralNullExpression(&'a mut LiteralNullExpression),
    LazyArrowExpressionWithFunctionBody(&'a mut LazyArrowExpressionWithFunctionBody),
    YieldStarExpression(&'a mut YieldStarExpression),
    CallExpression(&'a mut CallExpression),
    EagerArrowExpressionWithFunctionBody(&'a mut EagerArrowExpressionWithFunctionBody),
    UpdateExpression(&'a mut UpdateExpression),
    NewExpression(&'a mut NewExpression),
    IdentifierExpression(&'a mut IdentifierExpression),
    AssignmentExpression(&'a mut AssignmentExpression),
    UnaryExpression(&'a mut UnaryExpression),
    LiteralNumericExpression(&'a mut LiteralNumericExpression),
}

A mechanism to view value as an instance of interface sum ExpressionOrTemplateElement

Used to perform shallow cast between larger sums and smaller sums.

Variants

LiteralBooleanExpression(&'a mut LiteralBooleanExpression)

ObjectExpression(&'a mut ObjectExpression)

ComputedMemberExpression(&'a mut ComputedMemberExpression)

NewTargetExpression(&'a mut NewTargetExpression)

AwaitExpression(&'a mut AwaitExpression)

StaticMemberExpression(&'a mut StaticMemberExpression)

LazyArrowExpressionWithExpression(&'a mut LazyArrowExpressionWithExpression)

ThisExpression(&'a mut ThisExpression)

EagerArrowExpressionWithExpression(&'a mut EagerArrowExpressionWithExpression)

LazyFunctionExpression(&'a mut LazyFunctionExpression)

LiteralRegExpExpression(&'a mut LiteralRegExpExpression)

ConditionalExpression(&'a mut ConditionalExpression)

BinaryExpression(&'a mut BinaryExpression)

BinASTExpressionWithProbabilityTable(&'a mut BinASTExpressionWithProbabilityTable)

YieldExpression(&'a mut YieldExpression)

ArrayExpression(&'a mut ArrayExpression)

ClassExpression(&'a mut ClassExpression)

TemplateExpression(&'a mut TemplateExpression)

EagerFunctionExpression(&'a mut EagerFunctionExpression)

LiteralInfinityExpression(&'a mut LiteralInfinityExpression)

LiteralStringExpression(&'a mut LiteralStringExpression)

CompoundAssignmentExpression(&'a mut CompoundAssignmentExpression)

TemplateElement(&'a mut TemplateElement)

LiteralNullExpression(&'a mut LiteralNullExpression)

LazyArrowExpressionWithFunctionBody(&'a mut LazyArrowExpressionWithFunctionBody)

YieldStarExpression(&'a mut YieldStarExpression)

CallExpression(&'a mut CallExpression)

EagerArrowExpressionWithFunctionBody(&'a mut EagerArrowExpressionWithFunctionBody)

UpdateExpression(&'a mut UpdateExpression)

NewExpression(&'a mut NewExpression)

IdentifierExpression(&'a mut IdentifierExpression)

AssignmentExpression(&'a mut AssignmentExpression)

UnaryExpression(&'a mut UnaryExpression)

LiteralNumericExpression(&'a mut LiteralNumericExpression)

Methods

impl<'a> ViewMutExpressionOrTemplateElement<'a>

pub fn steal(&mut self) -> ExpressionOrTemplateElement

Convert a ViewMutExpressionOrTemplateElement back into a ExpressionOrTemplateElement, stealing the contents of the sum.