Enum binjs_es6::ast::ViewMutSpreadElementOrExpression

pub enum ViewMutSpreadElementOrExpression<'a> {
    EagerFunctionExpression(&'a mut EagerFunctionExpression),
    TemplateExpression(&'a mut TemplateExpression),
    LiteralStringExpression(&'a mut LiteralStringExpression),
    CallExpression(&'a mut CallExpression),
    EagerArrowExpressionWithFunctionBody(&'a mut EagerArrowExpressionWithFunctionBody),
    UpdateExpression(&'a mut UpdateExpression),
    LiteralNullExpression(&'a mut LiteralNullExpression),
    SpreadElement(&'a mut SpreadElement),
    ObjectExpression(&'a mut ObjectExpression),
    AssignmentExpression(&'a mut AssignmentExpression),
    CompoundAssignmentExpression(&'a mut CompoundAssignmentExpression),
    YieldExpression(&'a mut YieldExpression),
    ArrayExpression(&'a mut ArrayExpression),
    ConditionalExpression(&'a mut ConditionalExpression),
    EagerArrowExpressionWithExpression(&'a mut EagerArrowExpressionWithExpression),
    ThisExpression(&'a mut ThisExpression),
    LiteralInfinityExpression(&'a mut LiteralInfinityExpression),
    NewTargetExpression(&'a mut NewTargetExpression),
    LiteralNumericExpression(&'a mut LiteralNumericExpression),
    ClassExpression(&'a mut ClassExpression),
    BinaryExpression(&'a mut BinaryExpression),
    AwaitExpression(&'a mut AwaitExpression),
    YieldStarExpression(&'a mut YieldStarExpression),
    LiteralBooleanExpression(&'a mut LiteralBooleanExpression),
    LiteralRegExpExpression(&'a mut LiteralRegExpExpression),
    NewExpression(&'a mut NewExpression),
    ComputedMemberExpression(&'a mut ComputedMemberExpression),
    UnaryExpression(&'a mut UnaryExpression),
    LazyFunctionExpression(&'a mut LazyFunctionExpression),
    LazyArrowExpressionWithFunctionBody(&'a mut LazyArrowExpressionWithFunctionBody),
    IdentifierExpression(&'a mut IdentifierExpression),
    LazyArrowExpressionWithExpression(&'a mut LazyArrowExpressionWithExpression),
    StaticMemberExpression(&'a mut StaticMemberExpression),
}

A mechanism to view value as an instance of interface sum SpreadElementOrExpression

Used to perform shallow cast between larger sums and smaller sums.

Variants
EagerFunctionExpression(&'a mut EagerFunctionExpression)TemplateExpression(&'a mut TemplateExpression)LiteralStringExpression(&'a mut LiteralStringExpression)CallExpression(&'a mut CallExpression)EagerArrowExpressionWithFunctionBody(&'a mut EagerArrowExpressionWithFunctionBody)UpdateExpression(&'a mut UpdateExpression)LiteralNullExpression(&'a mut LiteralNullExpression)SpreadElement(&'a mut SpreadElement)ObjectExpression(&'a mut ObjectExpression)AssignmentExpression(&'a mut AssignmentExpression)CompoundAssignmentExpression(&'a mut CompoundAssignmentExpression)YieldExpression(&'a mut YieldExpression)ArrayExpression(&'a mut ArrayExpression)ConditionalExpression(&'a mut ConditionalExpression)EagerArrowExpressionWithExpression(&'a mut EagerArrowExpressionWithExpression)ThisExpression(&'a mut ThisExpression)LiteralInfinityExpression(&'a mut LiteralInfinityExpression)NewTargetExpression(&'a mut NewTargetExpression)LiteralNumericExpression(&'a mut LiteralNumericExpression)ClassExpression(&'a mut ClassExpression)BinaryExpression(&'a mut BinaryExpression)AwaitExpression(&'a mut AwaitExpression)YieldStarExpression(&'a mut YieldStarExpression)LiteralBooleanExpression(&'a mut LiteralBooleanExpression)LiteralRegExpExpression(&'a mut LiteralRegExpExpression)NewExpression(&'a mut NewExpression)ComputedMemberExpression(&'a mut ComputedMemberExpression)UnaryExpression(&'a mut UnaryExpression)LazyFunctionExpression(&'a mut LazyFunctionExpression)LazyArrowExpressionWithFunctionBody(&'a mut LazyArrowExpressionWithFunctionBody)IdentifierExpression(&'a mut IdentifierExpression)LazyArrowExpressionWithExpression(&'a mut LazyArrowExpressionWithExpression)StaticMemberExpression(&'a mut StaticMemberExpression)