Trait binjs_es6::ast::Visitor

pub trait Visitor<E> {
    fn enter_null(&mut self, _path: &Path, _node: &mut Null) -> Result<(), E> { ... }

    fn exit_null(&mut self, _path: &Path, _node: &mut Null) -> Result<(), E> { ... }

    fn enter_array_assignment_target(
        &mut self, 
        _path: &Path, 
        _node: &mut ArrayAssignmentTarget
    ) -> Result<(), E> { ... }

    fn exit_array_assignment_target(
        &mut self, 
        _path: &Path, 
        _node: &mut ArrayAssignmentTarget
    ) -> Result<(), E> { ... }

    fn enter_array_binding(
        &mut self, 
        _path: &Path, 
        _node: &mut ArrayBinding
    ) -> Result<(), E> { ... }

    fn exit_array_binding(
        &mut self, 
        _path: &Path, 
        _node: &mut ArrayBinding
    ) -> Result<(), E> { ... }

    fn enter_array_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ArrayExpression
    ) -> Result<(), E> { ... }

    fn exit_array_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ArrayExpression
    ) -> Result<(), E> { ... }

    fn enter_asserted_block_scope(
        &mut self, 
        _path: &Path, 
        _node: &mut AssertedBlockScope
    ) -> Result<(), E> { ... }

    fn exit_asserted_block_scope(
        &mut self, 
        _path: &Path, 
        _node: &mut AssertedBlockScope
    ) -> Result<(), E> { ... }

    fn enter_asserted_parameter_scope(
        &mut self, 
        _path: &Path, 
        _node: &mut AssertedParameterScope
    ) -> Result<(), E> { ... }

    fn exit_asserted_parameter_scope(
        &mut self, 
        _path: &Path, 
        _node: &mut AssertedParameterScope
    ) -> Result<(), E> { ... }

    fn enter_asserted_var_scope(
        &mut self, 
        _path: &Path, 
        _node: &mut AssertedVarScope
    ) -> Result<(), E> { ... }

    fn exit_asserted_var_scope(
        &mut self, 
        _path: &Path, 
        _node: &mut AssertedVarScope
    ) -> Result<(), E> { ... }

    fn enter_assignment_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentExpression
    ) -> Result<(), E> { ... }

    fn exit_assignment_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentExpression
    ) -> Result<(), E> { ... }

    fn enter_assignment_target_identifier(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentTargetIdentifier
    ) -> Result<(), E> { ... }

    fn exit_assignment_target_identifier(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentTargetIdentifier
    ) -> Result<(), E> { ... }

    fn enter_assignment_target_property_identifier(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentTargetPropertyIdentifier
    ) -> Result<(), E> { ... }

    fn exit_assignment_target_property_identifier(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentTargetPropertyIdentifier
    ) -> Result<(), E> { ... }

    fn enter_assignment_target_property_property(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentTargetPropertyProperty
    ) -> Result<(), E> { ... }

    fn exit_assignment_target_property_property(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentTargetPropertyProperty
    ) -> Result<(), E> { ... }

    fn enter_assignment_target_with_initializer(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentTargetWithInitializer
    ) -> Result<(), E> { ... }

    fn exit_assignment_target_with_initializer(
        &mut self, 
        _path: &Path, 
        _node: &mut AssignmentTargetWithInitializer
    ) -> Result<(), E> { ... }

    fn enter_await_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut AwaitExpression
    ) -> Result<(), E> { ... }

    fn exit_await_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut AwaitExpression
    ) -> Result<(), E> { ... }

    fn enter_binary_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut BinaryExpression
    ) -> Result<(), E> { ... }

    fn exit_binary_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut BinaryExpression
    ) -> Result<(), E> { ... }

    fn enter_binding_identifier(
        &mut self, 
        _path: &Path, 
        _node: &mut BindingIdentifier
    ) -> Result<(), E> { ... }

    fn exit_binding_identifier(
        &mut self, 
        _path: &Path, 
        _node: &mut BindingIdentifier
    ) -> Result<(), E> { ... }

    fn enter_binding_property_identifier(
        &mut self, 
        _path: &Path, 
        _node: &mut BindingPropertyIdentifier
    ) -> Result<(), E> { ... }

    fn exit_binding_property_identifier(
        &mut self, 
        _path: &Path, 
        _node: &mut BindingPropertyIdentifier
    ) -> Result<(), E> { ... }

    fn enter_binding_property_property(
        &mut self, 
        _path: &Path, 
        _node: &mut BindingPropertyProperty
    ) -> Result<(), E> { ... }

    fn exit_binding_property_property(
        &mut self, 
        _path: &Path, 
        _node: &mut BindingPropertyProperty
    ) -> Result<(), E> { ... }

    fn enter_binding_with_initializer(
        &mut self, 
        _path: &Path, 
        _node: &mut BindingWithInitializer
    ) -> Result<(), E> { ... }

    fn exit_binding_with_initializer(
        &mut self, 
        _path: &Path, 
        _node: &mut BindingWithInitializer
    ) -> Result<(), E> { ... }

    fn enter_block(&mut self, _path: &Path, _node: &mut Block) -> Result<(), E> { ... }

    fn exit_block(&mut self, _path: &Path, _node: &mut Block) -> Result<(), E> { ... }

    fn enter_break_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut BreakStatement
    ) -> Result<(), E> { ... }

    fn exit_break_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut BreakStatement
    ) -> Result<(), E> { ... }

    fn enter_call_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut CallExpression
    ) -> Result<(), E> { ... }

    fn exit_call_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut CallExpression
    ) -> Result<(), E> { ... }

    fn enter_catch_clause(
        &mut self, 
        _path: &Path, 
        _node: &mut CatchClause
    ) -> Result<(), E> { ... }

    fn exit_catch_clause(
        &mut self, 
        _path: &Path, 
        _node: &mut CatchClause
    ) -> Result<(), E> { ... }

    fn enter_class_declaration(
        &mut self, 
        _path: &Path, 
        _node: &mut ClassDeclaration
    ) -> Result<(), E> { ... }

    fn exit_class_declaration(
        &mut self, 
        _path: &Path, 
        _node: &mut ClassDeclaration
    ) -> Result<(), E> { ... }

    fn enter_class_element(
        &mut self, 
        _path: &Path, 
        _node: &mut ClassElement
    ) -> Result<(), E> { ... }

    fn exit_class_element(
        &mut self, 
        _path: &Path, 
        _node: &mut ClassElement
    ) -> Result<(), E> { ... }

    fn enter_class_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ClassExpression
    ) -> Result<(), E> { ... }

    fn exit_class_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ClassExpression
    ) -> Result<(), E> { ... }

    fn enter_compound_assignment_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut CompoundAssignmentExpression
    ) -> Result<(), E> { ... }

    fn exit_compound_assignment_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut CompoundAssignmentExpression
    ) -> Result<(), E> { ... }

    fn enter_computed_member_assignment_target(
        &mut self, 
        _path: &Path, 
        _node: &mut ComputedMemberAssignmentTarget
    ) -> Result<(), E> { ... }

    fn exit_computed_member_assignment_target(
        &mut self, 
        _path: &Path, 
        _node: &mut ComputedMemberAssignmentTarget
    ) -> Result<(), E> { ... }

    fn enter_computed_member_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ComputedMemberExpression
    ) -> Result<(), E> { ... }

    fn exit_computed_member_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ComputedMemberExpression
    ) -> Result<(), E> { ... }

    fn enter_computed_property_name(
        &mut self, 
        _path: &Path, 
        _node: &mut ComputedPropertyName
    ) -> Result<(), E> { ... }

    fn exit_computed_property_name(
        &mut self, 
        _path: &Path, 
        _node: &mut ComputedPropertyName
    ) -> Result<(), E> { ... }

    fn enter_conditional_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ConditionalExpression
    ) -> Result<(), E> { ... }

    fn exit_conditional_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ConditionalExpression
    ) -> Result<(), E> { ... }

    fn enter_continue_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ContinueStatement
    ) -> Result<(), E> { ... }

    fn exit_continue_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ContinueStatement
    ) -> Result<(), E> { ... }

    fn enter_data_property(
        &mut self, 
        _path: &Path, 
        _node: &mut DataProperty
    ) -> Result<(), E> { ... }

    fn exit_data_property(
        &mut self, 
        _path: &Path, 
        _node: &mut DataProperty
    ) -> Result<(), E> { ... }

    fn enter_debugger_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut DebuggerStatement
    ) -> Result<(), E> { ... }

    fn exit_debugger_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut DebuggerStatement
    ) -> Result<(), E> { ... }

    fn enter_directive(
        &mut self, 
        _path: &Path, 
        _node: &mut Directive
    ) -> Result<(), E> { ... }

    fn exit_directive(
        &mut self, 
        _path: &Path, 
        _node: &mut Directive
    ) -> Result<(), E> { ... }

    fn enter_do_while_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut DoWhileStatement
    ) -> Result<(), E> { ... }

    fn exit_do_while_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut DoWhileStatement
    ) -> Result<(), E> { ... }

    fn enter_eager_arrow_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerArrowExpression
    ) -> Result<(), E> { ... }

    fn exit_eager_arrow_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerArrowExpression
    ) -> Result<(), E> { ... }

    fn enter_eager_function_declaration(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerFunctionDeclaration
    ) -> Result<(), E> { ... }

    fn exit_eager_function_declaration(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerFunctionDeclaration
    ) -> Result<(), E> { ... }

    fn enter_eager_function_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerFunctionExpression
    ) -> Result<(), E> { ... }

    fn exit_eager_function_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerFunctionExpression
    ) -> Result<(), E> { ... }

    fn enter_eager_getter(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerGetter
    ) -> Result<(), E> { ... }

    fn exit_eager_getter(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerGetter
    ) -> Result<(), E> { ... }

    fn enter_eager_method(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerMethod
    ) -> Result<(), E> { ... }

    fn exit_eager_method(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerMethod
    ) -> Result<(), E> { ... }

    fn enter_eager_setter(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerSetter
    ) -> Result<(), E> { ... }

    fn exit_eager_setter(
        &mut self, 
        _path: &Path, 
        _node: &mut EagerSetter
    ) -> Result<(), E> { ... }

    fn enter_empty_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut EmptyStatement
    ) -> Result<(), E> { ... }

    fn exit_empty_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut EmptyStatement
    ) -> Result<(), E> { ... }

    fn enter_export(
        &mut self, 
        _path: &Path, 
        _node: &mut Export
    ) -> Result<(), E> { ... }

    fn exit_export(&mut self, _path: &Path, _node: &mut Export) -> Result<(), E> { ... }

    fn enter_export_all_from(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportAllFrom
    ) -> Result<(), E> { ... }

    fn exit_export_all_from(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportAllFrom
    ) -> Result<(), E> { ... }

    fn enter_export_default(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportDefault
    ) -> Result<(), E> { ... }

    fn exit_export_default(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportDefault
    ) -> Result<(), E> { ... }

    fn enter_export_from(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportFrom
    ) -> Result<(), E> { ... }

    fn exit_export_from(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportFrom
    ) -> Result<(), E> { ... }

    fn enter_export_from_specifier(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportFromSpecifier
    ) -> Result<(), E> { ... }

    fn exit_export_from_specifier(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportFromSpecifier
    ) -> Result<(), E> { ... }

    fn enter_export_local_specifier(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportLocalSpecifier
    ) -> Result<(), E> { ... }

    fn exit_export_local_specifier(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportLocalSpecifier
    ) -> Result<(), E> { ... }

    fn enter_export_locals(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportLocals
    ) -> Result<(), E> { ... }

    fn exit_export_locals(
        &mut self, 
        _path: &Path, 
        _node: &mut ExportLocals
    ) -> Result<(), E> { ... }

    fn enter_expression_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ExpressionStatement
    ) -> Result<(), E> { ... }

    fn exit_expression_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ExpressionStatement
    ) -> Result<(), E> { ... }

    fn enter_for_in_of_binding(
        &mut self, 
        _path: &Path, 
        _node: &mut ForInOfBinding
    ) -> Result<(), E> { ... }

    fn exit_for_in_of_binding(
        &mut self, 
        _path: &Path, 
        _node: &mut ForInOfBinding
    ) -> Result<(), E> { ... }

    fn enter_for_in_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ForInStatement
    ) -> Result<(), E> { ... }

    fn exit_for_in_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ForInStatement
    ) -> Result<(), E> { ... }

    fn enter_for_of_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ForOfStatement
    ) -> Result<(), E> { ... }

    fn exit_for_of_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ForOfStatement
    ) -> Result<(), E> { ... }

    fn enter_for_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ForStatement
    ) -> Result<(), E> { ... }

    fn exit_for_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ForStatement
    ) -> Result<(), E> { ... }

    fn enter_formal_parameters(
        &mut self, 
        _path: &Path, 
        _node: &mut FormalParameters
    ) -> Result<(), E> { ... }

    fn exit_formal_parameters(
        &mut self, 
        _path: &Path, 
        _node: &mut FormalParameters
    ) -> Result<(), E> { ... }

    fn enter_function_body(
        &mut self, 
        _path: &Path, 
        _node: &mut FunctionBody
    ) -> Result<(), E> { ... }

    fn exit_function_body(
        &mut self, 
        _path: &Path, 
        _node: &mut FunctionBody
    ) -> Result<(), E> { ... }

    fn enter_identifier_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut IdentifierExpression
    ) -> Result<(), E> { ... }

    fn exit_identifier_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut IdentifierExpression
    ) -> Result<(), E> { ... }

    fn enter_if_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut IfStatement
    ) -> Result<(), E> { ... }

    fn exit_if_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut IfStatement
    ) -> Result<(), E> { ... }

    fn enter_import(
        &mut self, 
        _path: &Path, 
        _node: &mut Import
    ) -> Result<(), E> { ... }

    fn exit_import(&mut self, _path: &Path, _node: &mut Import) -> Result<(), E> { ... }

    fn enter_import_namespace(
        &mut self, 
        _path: &Path, 
        _node: &mut ImportNamespace
    ) -> Result<(), E> { ... }

    fn exit_import_namespace(
        &mut self, 
        _path: &Path, 
        _node: &mut ImportNamespace
    ) -> Result<(), E> { ... }

    fn enter_import_specifier(
        &mut self, 
        _path: &Path, 
        _node: &mut ImportSpecifier
    ) -> Result<(), E> { ... }

    fn exit_import_specifier(
        &mut self, 
        _path: &Path, 
        _node: &mut ImportSpecifier
    ) -> Result<(), E> { ... }

    fn enter_labelled_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut LabelledStatement
    ) -> Result<(), E> { ... }

    fn exit_labelled_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut LabelledStatement
    ) -> Result<(), E> { ... }

    fn enter_literal_boolean_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralBooleanExpression
    ) -> Result<(), E> { ... }

    fn exit_literal_boolean_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralBooleanExpression
    ) -> Result<(), E> { ... }

    fn enter_literal_infinity_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralInfinityExpression
    ) -> Result<(), E> { ... }

    fn exit_literal_infinity_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralInfinityExpression
    ) -> Result<(), E> { ... }

    fn enter_literal_null_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralNullExpression
    ) -> Result<(), E> { ... }

    fn exit_literal_null_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralNullExpression
    ) -> Result<(), E> { ... }

    fn enter_literal_numeric_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralNumericExpression
    ) -> Result<(), E> { ... }

    fn exit_literal_numeric_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralNumericExpression
    ) -> Result<(), E> { ... }

    fn enter_literal_property_name(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralPropertyName
    ) -> Result<(), E> { ... }

    fn exit_literal_property_name(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralPropertyName
    ) -> Result<(), E> { ... }

    fn enter_literal_reg_exp_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralRegExpExpression
    ) -> Result<(), E> { ... }

    fn exit_literal_reg_exp_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralRegExpExpression
    ) -> Result<(), E> { ... }

    fn enter_literal_string_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralStringExpression
    ) -> Result<(), E> { ... }

    fn exit_literal_string_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut LiteralStringExpression
    ) -> Result<(), E> { ... }

    fn enter_module(
        &mut self, 
        _path: &Path, 
        _node: &mut Module
    ) -> Result<(), E> { ... }

    fn exit_module(&mut self, _path: &Path, _node: &mut Module) -> Result<(), E> { ... }

    fn enter_new_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut NewExpression
    ) -> Result<(), E> { ... }

    fn exit_new_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut NewExpression
    ) -> Result<(), E> { ... }

    fn enter_new_target_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut NewTargetExpression
    ) -> Result<(), E> { ... }

    fn exit_new_target_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut NewTargetExpression
    ) -> Result<(), E> { ... }

    fn enter_object_assignment_target(
        &mut self, 
        _path: &Path, 
        _node: &mut ObjectAssignmentTarget
    ) -> Result<(), E> { ... }

    fn exit_object_assignment_target(
        &mut self, 
        _path: &Path, 
        _node: &mut ObjectAssignmentTarget
    ) -> Result<(), E> { ... }

    fn enter_object_binding(
        &mut self, 
        _path: &Path, 
        _node: &mut ObjectBinding
    ) -> Result<(), E> { ... }

    fn exit_object_binding(
        &mut self, 
        _path: &Path, 
        _node: &mut ObjectBinding
    ) -> Result<(), E> { ... }

    fn enter_object_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ObjectExpression
    ) -> Result<(), E> { ... }

    fn exit_object_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ObjectExpression
    ) -> Result<(), E> { ... }

    fn enter_return_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ReturnStatement
    ) -> Result<(), E> { ... }

    fn exit_return_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ReturnStatement
    ) -> Result<(), E> { ... }

    fn enter_script(
        &mut self, 
        _path: &Path, 
        _node: &mut Script
    ) -> Result<(), E> { ... }

    fn exit_script(&mut self, _path: &Path, _node: &mut Script) -> Result<(), E> { ... }

    fn enter_shorthand_property(
        &mut self, 
        _path: &Path, 
        _node: &mut ShorthandProperty
    ) -> Result<(), E> { ... }

    fn exit_shorthand_property(
        &mut self, 
        _path: &Path, 
        _node: &mut ShorthandProperty
    ) -> Result<(), E> { ... }

    fn enter_skippable_arrow_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableArrowExpression
    ) -> Result<(), E> { ... }

    fn exit_skippable_arrow_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableArrowExpression
    ) -> Result<(), E> { ... }

    fn enter_skippable_function_declaration(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableFunctionDeclaration
    ) -> Result<(), E> { ... }

    fn exit_skippable_function_declaration(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableFunctionDeclaration
    ) -> Result<(), E> { ... }

    fn enter_skippable_function_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableFunctionExpression
    ) -> Result<(), E> { ... }

    fn exit_skippable_function_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableFunctionExpression
    ) -> Result<(), E> { ... }

    fn enter_skippable_getter(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableGetter
    ) -> Result<(), E> { ... }

    fn exit_skippable_getter(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableGetter
    ) -> Result<(), E> { ... }

    fn enter_skippable_method(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableMethod
    ) -> Result<(), E> { ... }

    fn exit_skippable_method(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableMethod
    ) -> Result<(), E> { ... }

    fn enter_skippable_setter(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableSetter
    ) -> Result<(), E> { ... }

    fn exit_skippable_setter(
        &mut self, 
        _path: &Path, 
        _node: &mut SkippableSetter
    ) -> Result<(), E> { ... }

    fn enter_spread_element(
        &mut self, 
        _path: &Path, 
        _node: &mut SpreadElement
    ) -> Result<(), E> { ... }

    fn exit_spread_element(
        &mut self, 
        _path: &Path, 
        _node: &mut SpreadElement
    ) -> Result<(), E> { ... }

    fn enter_static_member_assignment_target(
        &mut self, 
        _path: &Path, 
        _node: &mut StaticMemberAssignmentTarget
    ) -> Result<(), E> { ... }

    fn exit_static_member_assignment_target(
        &mut self, 
        _path: &Path, 
        _node: &mut StaticMemberAssignmentTarget
    ) -> Result<(), E> { ... }

    fn enter_static_member_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut StaticMemberExpression
    ) -> Result<(), E> { ... }

    fn exit_static_member_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut StaticMemberExpression
    ) -> Result<(), E> { ... }

    fn enter_super_(&mut self, _path: &Path, _node: &mut Super) -> Result<(), E> { ... }

    fn exit_super_(&mut self, _path: &Path, _node: &mut Super) -> Result<(), E> { ... }

    fn enter_switch_case(
        &mut self, 
        _path: &Path, 
        _node: &mut SwitchCase
    ) -> Result<(), E> { ... }

    fn exit_switch_case(
        &mut self, 
        _path: &Path, 
        _node: &mut SwitchCase
    ) -> Result<(), E> { ... }

    fn enter_switch_default(
        &mut self, 
        _path: &Path, 
        _node: &mut SwitchDefault
    ) -> Result<(), E> { ... }

    fn exit_switch_default(
        &mut self, 
        _path: &Path, 
        _node: &mut SwitchDefault
    ) -> Result<(), E> { ... }

    fn enter_switch_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut SwitchStatement
    ) -> Result<(), E> { ... }

    fn exit_switch_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut SwitchStatement
    ) -> Result<(), E> { ... }

    fn enter_switch_statement_with_default(
        &mut self, 
        _path: &Path, 
        _node: &mut SwitchStatementWithDefault
    ) -> Result<(), E> { ... }

    fn exit_switch_statement_with_default(
        &mut self, 
        _path: &Path, 
        _node: &mut SwitchStatementWithDefault
    ) -> Result<(), E> { ... }

    fn enter_template_element(
        &mut self, 
        _path: &Path, 
        _node: &mut TemplateElement
    ) -> Result<(), E> { ... }

    fn exit_template_element(
        &mut self, 
        _path: &Path, 
        _node: &mut TemplateElement
    ) -> Result<(), E> { ... }

    fn enter_template_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut TemplateExpression
    ) -> Result<(), E> { ... }

    fn exit_template_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut TemplateExpression
    ) -> Result<(), E> { ... }

    fn enter_this_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ThisExpression
    ) -> Result<(), E> { ... }

    fn exit_this_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut ThisExpression
    ) -> Result<(), E> { ... }

    fn enter_throw_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ThrowStatement
    ) -> Result<(), E> { ... }

    fn exit_throw_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut ThrowStatement
    ) -> Result<(), E> { ... }

    fn enter_try_catch_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut TryCatchStatement
    ) -> Result<(), E> { ... }

    fn exit_try_catch_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut TryCatchStatement
    ) -> Result<(), E> { ... }

    fn enter_try_finally_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut TryFinallyStatement
    ) -> Result<(), E> { ... }

    fn exit_try_finally_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut TryFinallyStatement
    ) -> Result<(), E> { ... }

    fn enter_unary_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut UnaryExpression
    ) -> Result<(), E> { ... }

    fn exit_unary_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut UnaryExpression
    ) -> Result<(), E> { ... }

    fn enter_update_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut UpdateExpression
    ) -> Result<(), E> { ... }

    fn exit_update_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut UpdateExpression
    ) -> Result<(), E> { ... }

    fn enter_variable_declaration(
        &mut self, 
        _path: &Path, 
        _node: &mut VariableDeclaration
    ) -> Result<(), E> { ... }

    fn exit_variable_declaration(
        &mut self, 
        _path: &Path, 
        _node: &mut VariableDeclaration
    ) -> Result<(), E> { ... }

    fn enter_variable_declarator(
        &mut self, 
        _path: &Path, 
        _node: &mut VariableDeclarator
    ) -> Result<(), E> { ... }

    fn exit_variable_declarator(
        &mut self, 
        _path: &Path, 
        _node: &mut VariableDeclarator
    ) -> Result<(), E> { ... }

    fn enter_while_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut WhileStatement
    ) -> Result<(), E> { ... }

    fn exit_while_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut WhileStatement
    ) -> Result<(), E> { ... }

    fn enter_with_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut WithStatement
    ) -> Result<(), E> { ... }

    fn exit_with_statement(
        &mut self, 
        _path: &Path, 
        _node: &mut WithStatement
    ) -> Result<(), E> { ... }

    fn enter_yield_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut YieldExpression
    ) -> Result<(), E> { ... }

    fn exit_yield_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut YieldExpression
    ) -> Result<(), E> { ... }

    fn enter_yield_star_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut YieldStarExpression
    ) -> Result<(), E> { ... }

    fn exit_yield_star_expression(
        &mut self, 
        _path: &Path, 
        _node: &mut YieldStarExpression
    ) -> Result<(), E> { ... }
}

A set of callbacks used to inspect the contents of an AST in a strongly-typed manner. For each node Foo, enter_foo() will be called before visiting the children, giving the opportunity to alter the node, and enter_foo() will be called after visiting the children, giving the opportunity to alter it further.

Each of the nodes of this AST implements Walker and may be visited recursively using Visitor.

Provided Methods

fn enter_null(&mut self, _path: &Path, _node: &mut Null) -> Result<(), E>

fn exit_null(&mut self, _path: &Path, _node: &mut Null) -> Result<(), E>

fn enter_array_assignment_target(
&mut self,
_path: &Path,
_node: &mut ArrayAssignmentTarget
) -> Result<(), E>

fn exit_array_assignment_target(
&mut self,
_path: &Path,
_node: &mut ArrayAssignmentTarget
) -> Result<(), E>

fn enter_array_binding(
&mut self,
_path: &Path,
_node: &mut ArrayBinding
) -> Result<(), E>

fn exit_array_binding(
&mut self,
_path: &Path,
_node: &mut ArrayBinding
) -> Result<(), E>

fn enter_array_expression(
&mut self,
_path: &Path,
_node: &mut ArrayExpression
) -> Result<(), E>

fn exit_array_expression(
&mut self,
_path: &Path,
_node: &mut ArrayExpression
) -> Result<(), E>

fn enter_asserted_block_scope(
&mut self,
_path: &Path,
_node: &mut AssertedBlockScope
) -> Result<(), E>

fn exit_asserted_block_scope(
&mut self,
_path: &Path,
_node: &mut AssertedBlockScope
) -> Result<(), E>

fn enter_asserted_parameter_scope(
&mut self,
_path: &Path,
_node: &mut AssertedParameterScope
) -> Result<(), E>

fn exit_asserted_parameter_scope(
&mut self,
_path: &Path,
_node: &mut AssertedParameterScope
) -> Result<(), E>

fn enter_asserted_var_scope(
&mut self,
_path: &Path,
_node: &mut AssertedVarScope
) -> Result<(), E>

fn exit_asserted_var_scope(
&mut self,
_path: &Path,
_node: &mut AssertedVarScope
) -> Result<(), E>

fn enter_assignment_expression(
&mut self,
_path: &Path,
_node: &mut AssignmentExpression
) -> Result<(), E>

fn exit_assignment_expression(
&mut self,
_path: &Path,
_node: &mut AssignmentExpression
) -> Result<(), E>

fn enter_assignment_target_identifier(
&mut self,
_path: &Path,
_node: &mut AssignmentTargetIdentifier
) -> Result<(), E>

fn exit_assignment_target_identifier(
&mut self,
_path: &Path,
_node: &mut AssignmentTargetIdentifier
) -> Result<(), E>

fn enter_assignment_target_property_identifier(
&mut self,
_path: &Path,
_node: &mut AssignmentTargetPropertyIdentifier
) -> Result<(), E>

fn exit_assignment_target_property_identifier(
&mut self,
_path: &Path,
_node: &mut AssignmentTargetPropertyIdentifier
) -> Result<(), E>

fn enter_assignment_target_property_property(
&mut self,
_path: &Path,
_node: &mut AssignmentTargetPropertyProperty
) -> Result<(), E>

fn exit_assignment_target_property_property(
&mut self,
_path: &Path,
_node: &mut AssignmentTargetPropertyProperty
) -> Result<(), E>

fn enter_assignment_target_with_initializer(
&mut self,
_path: &Path,
_node: &mut AssignmentTargetWithInitializer
) -> Result<(), E>

fn exit_assignment_target_with_initializer(
&mut self,
_path: &Path,
_node: &mut AssignmentTargetWithInitializer
) -> Result<(), E>

fn enter_await_expression(
&mut self,
_path: &Path,
_node: &mut AwaitExpression
) -> Result<(), E>

fn exit_await_expression(
&mut self,
_path: &Path,
_node: &mut AwaitExpression
) -> Result<(), E>

fn enter_binary_expression(
&mut self,
_path: &Path,
_node: &mut BinaryExpression
) -> Result<(), E>

fn exit_binary_expression(
&mut self,
_path: &Path,
_node: &mut BinaryExpression
) -> Result<(), E>

fn enter_binding_identifier(
&mut self,
_path: &Path,
_node: &mut BindingIdentifier
) -> Result<(), E>

fn exit_binding_identifier(
&mut self,
_path: &Path,
_node: &mut BindingIdentifier
) -> Result<(), E>

fn enter_binding_property_identifier(
&mut self,
_path: &Path,
_node: &mut BindingPropertyIdentifier
) -> Result<(), E>

fn exit_binding_property_identifier(
&mut self,
_path: &Path,
_node: &mut BindingPropertyIdentifier
) -> Result<(), E>

fn enter_binding_property_property(
&mut self,
_path: &Path,
_node: &mut BindingPropertyProperty
) -> Result<(), E>

fn exit_binding_property_property(
&mut self,
_path: &Path,
_node: &mut BindingPropertyProperty
) -> Result<(), E>

fn enter_binding_with_initializer(
&mut self,
_path: &Path,
_node: &mut BindingWithInitializer
) -> Result<(), E>

fn exit_binding_with_initializer(
&mut self,
_path: &Path,
_node: &mut BindingWithInitializer
) -> Result<(), E>

fn enter_block(&mut self, _path: &Path, _node: &mut Block) -> Result<(), E>

fn exit_block(&mut self, _path: &Path, _node: &mut Block) -> Result<(), E>

fn enter_break_statement(
&mut self,
_path: &Path,
_node: &mut BreakStatement
) -> Result<(), E>

fn exit_break_statement(
&mut self,
_path: &Path,
_node: &mut BreakStatement
) -> Result<(), E>

fn enter_call_expression(
&mut self,
_path: &Path,
_node: &mut CallExpression
) -> Result<(), E>

fn exit_call_expression(
&mut self,
_path: &Path,
_node: &mut CallExpression
) -> Result<(), E>

fn enter_catch_clause(
&mut self,
_path: &Path,
_node: &mut CatchClause
) -> Result<(), E>

fn exit_catch_clause(
&mut self,
_path: &Path,
_node: &mut CatchClause
) -> Result<(), E>

fn enter_class_declaration(
&mut self,
_path: &Path,
_node: &mut ClassDeclaration
) -> Result<(), E>

fn exit_class_declaration(
&mut self,
_path: &Path,
_node: &mut ClassDeclaration
) -> Result<(), E>

fn enter_class_element(
&mut self,
_path: &Path,
_node: &mut ClassElement
) -> Result<(), E>

fn exit_class_element(
&mut self,
_path: &Path,
_node: &mut ClassElement
) -> Result<(), E>

fn enter_class_expression(
&mut self,
_path: &Path,
_node: &mut ClassExpression
) -> Result<(), E>

fn exit_class_expression(
&mut self,
_path: &Path,
_node: &mut ClassExpression
) -> Result<(), E>

fn enter_compound_assignment_expression(
&mut self,
_path: &Path,
_node: &mut CompoundAssignmentExpression
) -> Result<(), E>

fn exit_compound_assignment_expression(
&mut self,
_path: &Path,
_node: &mut CompoundAssignmentExpression
) -> Result<(), E>

fn enter_computed_member_assignment_target(
&mut self,
_path: &Path,
_node: &mut ComputedMemberAssignmentTarget
) -> Result<(), E>

fn exit_computed_member_assignment_target(
&mut self,
_path: &Path,
_node: &mut ComputedMemberAssignmentTarget
) -> Result<(), E>

fn enter_computed_member_expression(
&mut self,
_path: &Path,
_node: &mut ComputedMemberExpression
) -> Result<(), E>

fn exit_computed_member_expression(
&mut self,
_path: &Path,
_node: &mut ComputedMemberExpression
) -> Result<(), E>

fn enter_computed_property_name(
&mut self,
_path: &Path,
_node: &mut ComputedPropertyName
) -> Result<(), E>

fn exit_computed_property_name(
&mut self,
_path: &Path,
_node: &mut ComputedPropertyName
) -> Result<(), E>

fn enter_conditional_expression(
&mut self,
_path: &Path,
_node: &mut ConditionalExpression
) -> Result<(), E>

fn exit_conditional_expression(
&mut self,
_path: &Path,
_node: &mut ConditionalExpression
) -> Result<(), E>

fn enter_continue_statement(
&mut self,
_path: &Path,
_node: &mut ContinueStatement
) -> Result<(), E>

fn exit_continue_statement(
&mut self,
_path: &Path,
_node: &mut ContinueStatement
) -> Result<(), E>

fn enter_data_property(
&mut self,
_path: &Path,
_node: &mut DataProperty
) -> Result<(), E>

fn exit_data_property(
&mut self,
_path: &Path,
_node: &mut DataProperty
) -> Result<(), E>

fn enter_debugger_statement(
&mut self,
_path: &Path,
_node: &mut DebuggerStatement
) -> Result<(), E>

fn exit_debugger_statement(
&mut self,
_path: &Path,
_node: &mut DebuggerStatement
) -> Result<(), E>

fn enter_directive(
&mut self,
_path: &Path,
_node: &mut Directive
) -> Result<(), E>

fn exit_directive(
&mut self,
_path: &Path,
_node: &mut Directive
) -> Result<(), E>

fn enter_do_while_statement(
&mut self,
_path: &Path,
_node: &mut DoWhileStatement
) -> Result<(), E>

fn exit_do_while_statement(
&mut self,
_path: &Path,
_node: &mut DoWhileStatement
) -> Result<(), E>

fn enter_eager_arrow_expression(
&mut self,
_path: &Path,
_node: &mut EagerArrowExpression
) -> Result<(), E>

fn exit_eager_arrow_expression(
&mut self,
_path: &Path,
_node: &mut EagerArrowExpression
) -> Result<(), E>

fn enter_eager_function_declaration(
&mut self,
_path: &Path,
_node: &mut EagerFunctionDeclaration
) -> Result<(), E>

fn exit_eager_function_declaration(
&mut self,
_path: &Path,
_node: &mut EagerFunctionDeclaration
) -> Result<(), E>

fn enter_eager_function_expression(
&mut self,
_path: &Path,
_node: &mut EagerFunctionExpression
) -> Result<(), E>

fn exit_eager_function_expression(
&mut self,
_path: &Path,
_node: &mut EagerFunctionExpression
) -> Result<(), E>

fn enter_eager_getter(
&mut self,
_path: &Path,
_node: &mut EagerGetter
) -> Result<(), E>

fn exit_eager_getter(
&mut self,
_path: &Path,
_node: &mut EagerGetter
) -> Result<(), E>

fn enter_eager_method(
&mut self,
_path: &Path,
_node: &mut EagerMethod
) -> Result<(), E>

fn exit_eager_method(
&mut self,
_path: &Path,
_node: &mut EagerMethod
) -> Result<(), E>

fn enter_eager_setter(
&mut self,
_path: &Path,
_node: &mut EagerSetter
) -> Result<(), E>

fn exit_eager_setter(
&mut self,
_path: &Path,
_node: &mut EagerSetter
) -> Result<(), E>

fn enter_empty_statement(
&mut self,
_path: &Path,
_node: &mut EmptyStatement
) -> Result<(), E>

fn exit_empty_statement(
&mut self,
_path: &Path,
_node: &mut EmptyStatement
) -> Result<(), E>

fn enter_export(&mut self, _path: &Path, _node: &mut Export) -> Result<(), E>

fn exit_export(&mut self, _path: &Path, _node: &mut Export) -> Result<(), E>

fn enter_export_all_from(
&mut self,
_path: &Path,
_node: &mut ExportAllFrom
) -> Result<(), E>

fn exit_export_all_from(
&mut self,
_path: &Path,
_node: &mut ExportAllFrom
) -> Result<(), E>

fn enter_export_default(
&mut self,
_path: &Path,
_node: &mut ExportDefault
) -> Result<(), E>

fn exit_export_default(
&mut self,
_path: &Path,
_node: &mut ExportDefault
) -> Result<(), E>

fn enter_export_from(
&mut self,
_path: &Path,
_node: &mut ExportFrom
) -> Result<(), E>

fn exit_export_from(
&mut self,
_path: &Path,
_node: &mut ExportFrom
) -> Result<(), E>

fn enter_export_from_specifier(
&mut self,
_path: &Path,
_node: &mut ExportFromSpecifier
) -> Result<(), E>

fn exit_export_from_specifier(
&mut self,
_path: &Path,
_node: &mut ExportFromSpecifier
) -> Result<(), E>

fn enter_export_local_specifier(
&mut self,
_path: &Path,
_node: &mut ExportLocalSpecifier
) -> Result<(), E>

fn exit_export_local_specifier(
&mut self,
_path: &Path,
_node: &mut ExportLocalSpecifier
) -> Result<(), E>

fn enter_export_locals(
&mut self,
_path: &Path,
_node: &mut ExportLocals
) -> Result<(), E>

fn exit_export_locals(
&mut self,
_path: &Path,
_node: &mut ExportLocals
) -> Result<(), E>

fn enter_expression_statement(
&mut self,
_path: &Path,
_node: &mut ExpressionStatement
) -> Result<(), E>

fn exit_expression_statement(
&mut self,
_path: &Path,
_node: &mut ExpressionStatement
) -> Result<(), E>

fn enter_for_in_of_binding(
&mut self,
_path: &Path,
_node: &mut ForInOfBinding
) -> Result<(), E>

fn exit_for_in_of_binding(
&mut self,
_path: &Path,
_node: &mut ForInOfBinding
) -> Result<(), E>

fn enter_for_in_statement(
&mut self,
_path: &Path,
_node: &mut ForInStatement
) -> Result<(), E>

fn exit_for_in_statement(
&mut self,
_path: &Path,
_node: &mut ForInStatement
) -> Result<(), E>

fn enter_for_of_statement(
&mut self,
_path: &Path,
_node: &mut ForOfStatement
) -> Result<(), E>

fn exit_for_of_statement(
&mut self,
_path: &Path,
_node: &mut ForOfStatement
) -> Result<(), E>

fn enter_for_statement(
&mut self,
_path: &Path,
_node: &mut ForStatement
) -> Result<(), E>

fn exit_for_statement(
&mut self,
_path: &Path,
_node: &mut ForStatement
) -> Result<(), E>

fn enter_formal_parameters(
&mut self,
_path: &Path,
_node: &mut FormalParameters
) -> Result<(), E>

fn exit_formal_parameters(
&mut self,
_path: &Path,
_node: &mut FormalParameters
) -> Result<(), E>

fn enter_function_body(
&mut self,
_path: &Path,
_node: &mut FunctionBody
) -> Result<(), E>

fn exit_function_body(
&mut self,
_path: &Path,
_node: &mut FunctionBody
) -> Result<(), E>

fn enter_identifier_expression(
&mut self,
_path: &Path,
_node: &mut IdentifierExpression
) -> Result<(), E>

fn exit_identifier_expression(
&mut self,
_path: &Path,
_node: &mut IdentifierExpression
) -> Result<(), E>

fn enter_if_statement(
&mut self,
_path: &Path,
_node: &mut IfStatement
) -> Result<(), E>

fn exit_if_statement(
&mut self,
_path: &Path,
_node: &mut IfStatement
) -> Result<(), E>

fn enter_import(&mut self, _path: &Path, _node: &mut Import) -> Result<(), E>

fn exit_import(&mut self, _path: &Path, _node: &mut Import) -> Result<(), E>

fn enter_import_namespace(
&mut self,
_path: &Path,
_node: &mut ImportNamespace
) -> Result<(), E>

fn exit_import_namespace(
&mut self,
_path: &Path,
_node: &mut ImportNamespace
) -> Result<(), E>

fn enter_import_specifier(
&mut self,
_path: &Path,
_node: &mut ImportSpecifier
) -> Result<(), E>

fn exit_import_specifier(
&mut self,
_path: &Path,
_node: &mut ImportSpecifier
) -> Result<(), E>

fn enter_labelled_statement(
&mut self,
_path: &Path,
_node: &mut LabelledStatement
) -> Result<(), E>

fn exit_labelled_statement(
&mut self,
_path: &Path,
_node: &mut LabelledStatement
) -> Result<(), E>

fn enter_literal_boolean_expression(
&mut self,
_path: &Path,
_node: &mut LiteralBooleanExpression
) -> Result<(), E>

fn exit_literal_boolean_expression(
&mut self,
_path: &Path,
_node: &mut LiteralBooleanExpression
) -> Result<(), E>

fn enter_literal_infinity_expression(
&mut self,
_path: &Path,
_node: &mut LiteralInfinityExpression
) -> Result<(), E>

fn exit_literal_infinity_expression(
&mut self,
_path: &Path,
_node: &mut LiteralInfinityExpression
) -> Result<(), E>

fn enter_literal_null_expression(
&mut self,
_path: &Path,
_node: &mut LiteralNullExpression
) -> Result<(), E>

fn exit_literal_null_expression(
&mut self,
_path: &Path,
_node: &mut LiteralNullExpression
) -> Result<(), E>

fn enter_literal_numeric_expression(
&mut self,
_path: &Path,
_node: &mut LiteralNumericExpression
) -> Result<(), E>

fn exit_literal_numeric_expression(
&mut self,
_path: &Path,
_node: &mut LiteralNumericExpression
) -> Result<(), E>

fn enter_literal_property_name(
&mut self,
_path: &Path,
_node: &mut LiteralPropertyName
) -> Result<(), E>

fn exit_literal_property_name(
&mut self,
_path: &Path,
_node: &mut LiteralPropertyName
) -> Result<(), E>

fn enter_literal_reg_exp_expression(
&mut self,
_path: &Path,
_node: &mut LiteralRegExpExpression
) -> Result<(), E>

fn exit_literal_reg_exp_expression(
&mut self,
_path: &Path,
_node: &mut LiteralRegExpExpression
) -> Result<(), E>

fn enter_literal_string_expression(
&mut self,
_path: &Path,
_node: &mut LiteralStringExpression
) -> Result<(), E>

fn exit_literal_string_expression(
&mut self,
_path: &Path,
_node: &mut LiteralStringExpression
) -> Result<(), E>

fn enter_module(&mut self, _path: &Path, _node: &mut Module) -> Result<(), E>

fn exit_module(&mut self, _path: &Path, _node: &mut Module) -> Result<(), E>

fn enter_new_expression(
&mut self,
_path: &Path,
_node: &mut NewExpression
) -> Result<(), E>

fn exit_new_expression(
&mut self,
_path: &Path,
_node: &mut NewExpression
) -> Result<(), E>

fn enter_new_target_expression(
&mut self,
_path: &Path,
_node: &mut NewTargetExpression
) -> Result<(), E>

fn exit_new_target_expression(
&mut self,
_path: &Path,
_node: &mut NewTargetExpression
) -> Result<(), E>

fn enter_object_assignment_target(
&mut self,
_path: &Path,
_node: &mut ObjectAssignmentTarget
) -> Result<(), E>

fn exit_object_assignment_target(
&mut self,
_path: &Path,
_node: &mut ObjectAssignmentTarget
) -> Result<(), E>

fn enter_object_binding(
&mut self,
_path: &Path,
_node: &mut ObjectBinding
) -> Result<(), E>

fn exit_object_binding(
&mut self,
_path: &Path,
_node: &mut ObjectBinding
) -> Result<(), E>

fn enter_object_expression(
&mut self,
_path: &Path,
_node: &mut ObjectExpression
) -> Result<(), E>

fn exit_object_expression(
&mut self,
_path: &Path,
_node: &mut ObjectExpression
) -> Result<(), E>

fn enter_return_statement(
&mut self,
_path: &Path,
_node: &mut ReturnStatement
) -> Result<(), E>

fn exit_return_statement(
&mut self,
_path: &Path,
_node: &mut ReturnStatement
) -> Result<(), E>

fn enter_script(&mut self, _path: &Path, _node: &mut Script) -> Result<(), E>

fn exit_script(&mut self, _path: &Path, _node: &mut Script) -> Result<(), E>

fn enter_shorthand_property(
&mut self,
_path: &Path,
_node: &mut ShorthandProperty
) -> Result<(), E>

fn exit_shorthand_property(
&mut self,
_path: &Path,
_node: &mut ShorthandProperty
) -> Result<(), E>

fn enter_skippable_arrow_expression(
&mut self,
_path: &Path,
_node: &mut SkippableArrowExpression
) -> Result<(), E>

fn exit_skippable_arrow_expression(
&mut self,
_path: &Path,
_node: &mut SkippableArrowExpression
) -> Result<(), E>

fn enter_skippable_function_declaration(
&mut self,
_path: &Path,
_node: &mut SkippableFunctionDeclaration
) -> Result<(), E>

fn exit_skippable_function_declaration(
&mut self,
_path: &Path,
_node: &mut SkippableFunctionDeclaration
) -> Result<(), E>

fn enter_skippable_function_expression(
&mut self,
_path: &Path,
_node: &mut SkippableFunctionExpression
) -> Result<(), E>

fn exit_skippable_function_expression(
&mut self,
_path: &Path,
_node: &mut SkippableFunctionExpression
) -> Result<(), E>

fn enter_skippable_getter(
&mut self,
_path: &Path,
_node: &mut SkippableGetter
) -> Result<(), E>

fn exit_skippable_getter(
&mut self,
_path: &Path,
_node: &mut SkippableGetter
) -> Result<(), E>

fn enter_skippable_method(
&mut self,
_path: &Path,
_node: &mut SkippableMethod
) -> Result<(), E>

fn exit_skippable_method(
&mut self,
_path: &Path,
_node: &mut SkippableMethod
) -> Result<(), E>

fn enter_skippable_setter(
&mut self,
_path: &Path,
_node: &mut SkippableSetter
) -> Result<(), E>

fn exit_skippable_setter(
&mut self,
_path: &Path,
_node: &mut SkippableSetter
) -> Result<(), E>

fn enter_spread_element(
&mut self,
_path: &Path,
_node: &mut SpreadElement
) -> Result<(), E>

fn exit_spread_element(
&mut self,
_path: &Path,
_node: &mut SpreadElement
) -> Result<(), E>

fn enter_static_member_assignment_target(
&mut self,
_path: &Path,
_node: &mut StaticMemberAssignmentTarget
) -> Result<(), E>

fn exit_static_member_assignment_target(
&mut self,
_path: &Path,
_node: &mut StaticMemberAssignmentTarget
) -> Result<(), E>

fn enter_static_member_expression(
&mut self,
_path: &Path,
_node: &mut StaticMemberExpression
) -> Result<(), E>

fn exit_static_member_expression(
&mut self,
_path: &Path,
_node: &mut StaticMemberExpression
) -> Result<(), E>

fn enter_super_(&mut self, _path: &Path, _node: &mut Super) -> Result<(), E>

fn exit_super_(&mut self, _path: &Path, _node: &mut Super) -> Result<(), E>

fn enter_switch_case(
&mut self,
_path: &Path,
_node: &mut SwitchCase
) -> Result<(), E>

fn exit_switch_case(
&mut self,
_path: &Path,
_node: &mut SwitchCase
) -> Result<(), E>

fn enter_switch_default(
&mut self,
_path: &Path,
_node: &mut SwitchDefault
) -> Result<(), E>

fn exit_switch_default(
&mut self,
_path: &Path,
_node: &mut SwitchDefault
) -> Result<(), E>

fn enter_switch_statement(
&mut self,
_path: &Path,
_node: &mut SwitchStatement
) -> Result<(), E>

fn exit_switch_statement(
&mut self,
_path: &Path,
_node: &mut SwitchStatement
) -> Result<(), E>

fn enter_switch_statement_with_default(
&mut self,
_path: &Path,
_node: &mut SwitchStatementWithDefault
) -> Result<(), E>

fn exit_switch_statement_with_default(
&mut self,
_path: &Path,
_node: &mut SwitchStatementWithDefault
) -> Result<(), E>

fn enter_template_element(
&mut self,
_path: &Path,
_node: &mut TemplateElement
) -> Result<(), E>

fn exit_template_element(
&mut self,
_path: &Path,
_node: &mut TemplateElement
) -> Result<(), E>

fn enter_template_expression(
&mut self,
_path: &Path,
_node: &mut TemplateExpression
) -> Result<(), E>

fn exit_template_expression(
&mut self,
_path: &Path,
_node: &mut TemplateExpression
) -> Result<(), E>

fn enter_this_expression(
&mut self,
_path: &Path,
_node: &mut ThisExpression
) -> Result<(), E>

fn exit_this_expression(
&mut self,
_path: &Path,
_node: &mut ThisExpression
) -> Result<(), E>

fn enter_throw_statement(
&mut self,
_path: &Path,
_node: &mut ThrowStatement
) -> Result<(), E>

fn exit_throw_statement(
&mut self,
_path: &Path,
_node: &mut ThrowStatement
) -> Result<(), E>

fn enter_try_catch_statement(
&mut self,
_path: &Path,
_node: &mut TryCatchStatement
) -> Result<(), E>

fn exit_try_catch_statement(
&mut self,
_path: &Path,
_node: &mut TryCatchStatement
) -> Result<(), E>

fn enter_try_finally_statement(
&mut self,
_path: &Path,
_node: &mut TryFinallyStatement
) -> Result<(), E>

fn exit_try_finally_statement(
&mut self,
_path: &Path,
_node: &mut TryFinallyStatement
) -> Result<(), E>

fn enter_unary_expression(
&mut self,
_path: &Path,
_node: &mut UnaryExpression
) -> Result<(), E>

fn exit_unary_expression(
&mut self,
_path: &Path,
_node: &mut UnaryExpression
) -> Result<(), E>

fn enter_update_expression(
&mut self,
_path: &Path,
_node: &mut UpdateExpression
) -> Result<(), E>

fn exit_update_expression(
&mut self,
_path: &Path,
_node: &mut UpdateExpression
) -> Result<(), E>

fn enter_variable_declaration(
&mut self,
_path: &Path,
_node: &mut VariableDeclaration
) -> Result<(), E>

fn exit_variable_declaration(
&mut self,
_path: &Path,
_node: &mut VariableDeclaration
) -> Result<(), E>

fn enter_variable_declarator(
&mut self,
_path: &Path,
_node: &mut VariableDeclarator
) -> Result<(), E>

fn exit_variable_declarator(
&mut self,
_path: &Path,
_node: &mut VariableDeclarator
) -> Result<(), E>

fn enter_while_statement(
&mut self,
_path: &Path,
_node: &mut WhileStatement
) -> Result<(), E>

fn exit_while_statement(
&mut self,
_path: &Path,
_node: &mut WhileStatement
) -> Result<(), E>

fn enter_with_statement(
&mut self,
_path: &Path,
_node: &mut WithStatement
) -> Result<(), E>

fn exit_with_statement(
&mut self,
_path: &Path,
_node: &mut WithStatement
) -> Result<(), E>

fn enter_yield_expression(
&mut self,
_path: &Path,
_node: &mut YieldExpression
) -> Result<(), E>

fn exit_yield_expression(
&mut self,
_path: &Path,
_node: &mut YieldExpression
) -> Result<(), E>

fn enter_yield_star_expression(
&mut self,
_path: &Path,
_node: &mut YieldStarExpression
) -> Result<(), E>

fn exit_yield_star_expression(
&mut self,
_path: &Path,
_node: &mut YieldStarExpression
) -> Result<(), E>