Module binjs_es6::ast

A strongly-typed AST for ES6.

Structs

ArrayAssignmentTarget

ArrayBinding

ArrayExpression

AssertedBlockScope

AssertedParameterScope

AssertedVarScope

AssignmentExpression

AssignmentTargetIdentifier

AssignmentTargetPropertyIdentifier

AssignmentTargetPropertyProperty

AssignmentTargetWithInitializer

AwaitExpression

BinaryExpression

BindingIdentifier

BindingPropertyIdentifier

BindingPropertyProperty

BindingWithInitializer

Block

BreakStatement

CallExpression

CatchClause

ClassDeclaration

ClassElement

ClassExpression

CompoundAssignmentExpression

ComputedMemberAssignmentTarget

ComputedMemberExpression

ComputedPropertyName

ConditionalExpression

ContinueStatement

DataProperty

DebuggerStatement

Directive

DoWhileStatement

EagerArrowExpression

EagerFunctionDeclaration

EagerFunctionExpression

EagerGetter

EagerMethod

EagerSetter

EmptyStatement

Export

ExportAllFrom

ExportDefault

ExportFrom

ExportFromSpecifier

ExportLocalSpecifier

ExportLocals

ExpressionStatement

ForInOfBinding

ForInStatement

ForOfStatement

ForStatement

FormalParameters

FunctionBody

IdentifierExpression

IfStatement

Import

ImportNamespace

ImportSpecifier

LabelledStatement

LiteralBooleanExpression

LiteralInfinityExpression

LiteralNullExpression

LiteralNumericExpression

LiteralPropertyName

LiteralRegExpExpression

LiteralStringExpression

Module

NewExpression

NewTargetExpression

Null

ObjectAssignmentTarget

ObjectBinding

ObjectExpression

ReturnStatement

Script

ShorthandProperty

SkippableArrowExpression

SkippableFunctionDeclaration

SkippableFunctionExpression

SkippableGetter

SkippableMethod

SkippableSetter

SpreadElement

StaticMemberAssignmentTarget

StaticMemberExpression

Super

SwitchCase

SwitchDefault

SwitchStatement

SwitchStatementWithDefault

TemplateElement

TemplateExpression

ThisExpression

ThrowStatement

TryCatchStatement

TryFinallyStatement

UnaryExpression

UpdateExpression

VariableDeclaration

VariableDeclarator

ViewMutArguments

ViewMutArrayAssignmentTarget

ViewMutArrayBinding

ViewMutArrayExpression

ViewMutAssertedBlockScope

ViewMutAssertedParameterScope

ViewMutAssertedVarScope

ViewMutAssignmentExpression

ViewMutAssignmentTargetIdentifier

ViewMutAssignmentTargetPropertyIdentifier

ViewMutAssignmentTargetPropertyProperty

ViewMutAssignmentTargetWithInitializer

ViewMutAwaitExpression

ViewMutBinaryExpression

ViewMutBindingIdentifier

ViewMutBindingPropertyIdentifier

ViewMutBindingPropertyProperty

ViewMutBindingWithInitializer

ViewMutBlock

ViewMutBreakStatement

ViewMutCallExpression

ViewMutCatchClause

ViewMutClassDeclaration

ViewMutClassElement

ViewMutClassExpression

ViewMutCompoundAssignmentExpression

ViewMutComputedMemberAssignmentTarget

ViewMutComputedMemberExpression

ViewMutComputedPropertyName

ViewMutConditionalExpression

ViewMutContinueStatement

ViewMutDataProperty

ViewMutDebuggerStatement

ViewMutDirective

ViewMutDoWhileStatement

ViewMutEagerArrowExpression

ViewMutEagerFunctionDeclaration

ViewMutEagerFunctionExpression

ViewMutEagerGetter

ViewMutEagerMethod

ViewMutEagerSetter

ViewMutEmptyStatement

ViewMutExport

ViewMutExportAllFrom

ViewMutExportDefault

ViewMutExportFrom

ViewMutExportFromSpecifier

ViewMutExportLocalSpecifier

ViewMutExportLocals

ViewMutExpressionStatement

ViewMutForInOfBinding

ViewMutForInStatement

ViewMutForOfStatement

ViewMutForStatement

ViewMutFormalParameters

ViewMutFunctionBody

ViewMutIdentifier

ViewMutIdentifierExpression

ViewMutIdentifierName

ViewMutIfStatement

ViewMutImport

ViewMutImportNamespace

ViewMutImportSpecifier

ViewMutLabel

ViewMutLabelledStatement

ViewMutListOfAssignmentTargetOrAssignmentTargetWithInitializer

ViewMutListOfAssignmentTargetProperty

ViewMutListOfBindingProperty

ViewMutListOfClassElement

ViewMutListOfDirective

ViewMutListOfExportFromSpecifier

ViewMutListOfExportLocalSpecifier

ViewMutListOfExpressionOrTemplateElement

ViewMutListOfIdentifierName

ViewMutListOfImportDeclarationOrExportDeclarationOrStatement

ViewMutListOfImportSpecifier

ViewMutListOfObjectProperty

ViewMutListOfOptionalBindingOrBindingWithInitializer

ViewMutListOfOptionalSpreadElementOrExpression

ViewMutListOfParameter

ViewMutListOfStatement

ViewMutListOfSwitchCase

ViewMutListOfVariableDeclarator

ViewMutLiteralBooleanExpression

ViewMutLiteralInfinityExpression

ViewMutLiteralNullExpression

ViewMutLiteralNumericExpression

ViewMutLiteralPropertyName

ViewMutLiteralRegExpExpression

ViewMutLiteralStringExpression

ViewMutModule

ViewMutNewExpression

ViewMutNewTargetExpression

ViewMutNull

ViewMutObjectAssignmentTarget

ViewMutObjectBinding

ViewMutObjectExpression

ViewMutOffset

ViewMutOptionalAssertedBlockScope

ViewMutOptionalAssertedParameterScope

ViewMutOptionalAssertedVarScope

ViewMutOptionalAssignmentTarget

ViewMutOptionalBinding

ViewMutOptionalBindingIdentifier

ViewMutOptionalBindingOrBindingWithInitializer

ViewMutOptionalCatchClause

ViewMutOptionalExpression

ViewMutOptionalIdentifierName

ViewMutOptionalLabel

ViewMutOptionalSpreadElementOrExpression

ViewMutOptionalStatement

ViewMutOptionalVariableDeclarationOrExpression

ViewMutReturnStatement

ViewMutScript

ViewMutShorthandProperty

ViewMutSkippableArrowExpression

ViewMutSkippableFunctionDeclaration

ViewMutSkippableFunctionExpression

ViewMutSkippableGetter

ViewMutSkippableMethod

ViewMutSkippableSetter

ViewMutSpreadElement

ViewMutStaticMemberAssignmentTarget

ViewMutStaticMemberExpression

ViewMutString

ViewMutSuper

ViewMutSwitchCase

ViewMutSwitchDefault

ViewMutSwitchStatement

ViewMutSwitchStatementWithDefault

ViewMutTemplateElement

ViewMutTemplateExpression

ViewMutThisExpression

ViewMutThrowStatement

ViewMutTryCatchStatement

ViewMutTryFinallyStatement

ViewMutUnaryExpression

ViewMutUpdateExpression

ViewMutVariableDeclaration

ViewMutVariableDeclarator

ViewMutWhileStatement

ViewMutWithStatement

ViewMutYieldExpression

ViewMutYieldStarExpression

WhileStatement

WithStatement

YieldExpression

YieldStarExpression

Enums

ASTField

ASTNode

ArrowExpression

AssignmentTarget

AssignmentTargetOrAssignmentTargetWithInitializer

AssignmentTargetPattern

AssignmentTargetProperty

BinaryOperator

Binding

BindingOrBindingWithInitializer

BindingPattern

BindingProperty

CompoundAssignmentOperator

ExportDeclaration

Expression

ExpressionOrSuper

ExpressionOrTemplateElement

ForInOfBindingOrAssignmentTarget

FunctionBodyOrExpression

FunctionDeclaration

FunctionDeclarationOrClassDeclarationOrExpression

FunctionDeclarationOrClassDeclarationOrVariableDeclaration

FunctionExpression

Getter

ImportDeclaration

ImportDeclarationOrExportDeclarationOrStatement

IterationStatement

Literal

Method

MethodDefinition

ObjectProperty

Parameter

Program

PropertyName

Setter

SimpleAssignmentTarget

SpreadElementOrExpression

Statement

UnaryOperator

UpdateOperator

VariableDeclarationKind

VariableDeclarationOrExpression

ViewMutArrowExpression

ViewMutAssignmentTarget

ViewMutAssignmentTargetOrAssignmentTargetWithInitializer

ViewMutAssignmentTargetPattern

ViewMutAssignmentTargetProperty

ViewMutBinding

ViewMutBindingOrBindingWithInitializer

ViewMutBindingPattern

ViewMutBindingProperty

ViewMutExportDeclaration

ViewMutExpression

ViewMutExpressionOrSuper

ViewMutExpressionOrTemplateElement

ViewMutForInOfBindingOrAssignmentTarget

ViewMutFunctionBodyOrExpression

ViewMutFunctionDeclaration

ViewMutFunctionDeclarationOrClassDeclarationOrExpression

ViewMutFunctionDeclarationOrClassDeclarationOrVariableDeclaration

ViewMutFunctionExpression

ViewMutGetter

ViewMutImportDeclaration

ViewMutImportDeclarationOrExportDeclarationOrStatement

ViewMutIterationStatement

ViewMutLiteral

ViewMutMethod

ViewMutMethodDefinition

ViewMutObjectProperty

ViewMutParameter

ViewMutProgram

ViewMutPropertyName

ViewMutSetter

ViewMutSimpleAssignmentTarget

ViewMutSpreadElementOrExpression

ViewMutStatement

ViewMutVariableDeclarationOrExpression

Traits

Visitor
A set of callbacks used to inspect the contents of an AST in a strongly-typed manner. For each node Foo, enter_foo() will be called before visiting the children, giving the opportunity to alter the node, and enter_foo() will be called after visiting the children, giving the opportunity to alter it further.

Walker

Type Definitions

Arguments

Identifier

IdentifierName

Label

ListOfAssignmentTargetOrAssignmentTargetWithInitializer

ListOfAssignmentTargetProperty

ListOfBindingProperty

ListOfClassElement

ListOfDirective

ListOfExportFromSpecifier

ListOfExportLocalSpecifier

ListOfExpressionOrTemplateElement

ListOfIdentifierName

ListOfImportDeclarationOrExportDeclarationOrStatement

ListOfImportSpecifier

ListOfObjectProperty

ListOfOptionalBindingOrBindingWithInitializer

ListOfOptionalSpreadElementOrExpression

ListOfParameter

ListOfStatement

ListOfSwitchCase

ListOfVariableDeclarator

OptionalAssertedBlockScope

OptionalAssertedParameterScope

OptionalAssertedVarScope

OptionalAssignmentTarget

OptionalBinding

OptionalBindingIdentifier

OptionalBindingOrBindingWithInitializer

OptionalCatchClause

OptionalExpression

OptionalIdentifierName

OptionalLabel

OptionalSpreadElementOrExpression

OptionalStatement

OptionalVariableDeclarationOrExpression

Path

PathItem

String