Module binjs_es6::ast

A strongly-typed AST for ES6.

Structs

ArrayAssignmentTarget
Implementation of interface ArrayAssignmentTarget.

ArrayBinding
Implementation of interface ArrayBinding.

ArrayExpression
Implementation of interface ArrayExpression.

ArrowExpressionContentsWithExpression
Implementation of interface ArrowExpressionContentsWithExpression.

ArrowExpressionContentsWithFunctionBody
Implementation of interface ArrowExpressionContentsWithFunctionBody.

AssertedBlockScope
Implementation of interface AssertedBlockScope.

AssertedBoundName
Implementation of interface AssertedBoundName.

AssertedBoundNamesScope
Implementation of interface AssertedBoundNamesScope.

AssertedDeclaredName
Implementation of interface AssertedDeclaredName.

AssertedParameterName
Implementation of interface AssertedParameterName.

AssertedParameterScope
Implementation of interface AssertedParameterScope.

AssertedPositionalParameterName
Implementation of interface AssertedPositionalParameterName.

AssertedRestParameterName
Implementation of interface AssertedRestParameterName.

AssertedScriptGlobalScope
Implementation of interface AssertedScriptGlobalScope.

AssertedVarScope
Implementation of interface AssertedVarScope.

AssignmentExpression
Implementation of interface AssignmentExpression.

AssignmentTargetIdentifier
Implementation of interface AssignmentTargetIdentifier.

AssignmentTargetPropertyIdentifier
Implementation of interface AssignmentTargetPropertyIdentifier.

AssignmentTargetPropertyProperty
Implementation of interface AssignmentTargetPropertyProperty.

AssignmentTargetWithInitializer
Implementation of interface AssignmentTargetWithInitializer.

AwaitExpression
Implementation of interface AwaitExpression.

BinaryExpression
Implementation of interface BinaryExpression.

BindingIdentifier
Implementation of interface BindingIdentifier.

BindingPropertyIdentifier
Implementation of interface BindingPropertyIdentifier.

BindingPropertyProperty
Implementation of interface BindingPropertyProperty.

BindingWithInitializer
Implementation of interface BindingWithInitializer.

Block
Implementation of interface Block.

BreakStatement
Implementation of interface BreakStatement.

CallExpression
Implementation of interface CallExpression.

CatchClause
Implementation of interface CatchClause.

ClassDeclaration
Implementation of interface ClassDeclaration.

ClassElement
Implementation of interface ClassElement.

ClassExpression
Implementation of interface ClassExpression.

CompoundAssignmentExpression
Implementation of interface CompoundAssignmentExpression.

ComputedMemberAssignmentTarget
Implementation of interface ComputedMemberAssignmentTarget.

ComputedMemberExpression
Implementation of interface ComputedMemberExpression.

ComputedPropertyName
Implementation of interface ComputedPropertyName.

ConditionalExpression
Implementation of interface ConditionalExpression.

ContinueStatement
Implementation of interface ContinueStatement.

DataProperty
Implementation of interface DataProperty.

DebuggerStatement
Implementation of interface DebuggerStatement.

Directive
Implementation of interface Directive.

DoWhileStatement
Implementation of interface DoWhileStatement.

EagerArrowExpressionWithExpression
Implementation of interface EagerArrowExpressionWithExpression.

EagerArrowExpressionWithFunctionBody
Implementation of interface EagerArrowExpressionWithFunctionBody.

EagerFunctionDeclaration
Implementation of interface EagerFunctionDeclaration.

EagerFunctionExpression
Implementation of interface EagerFunctionExpression.

EagerGetter
Implementation of interface EagerGetter.

EagerMethod
Implementation of interface EagerMethod.

EagerSetter
Implementation of interface EagerSetter.

EmptyStatement
Implementation of interface EmptyStatement.

Export
Implementation of interface Export.

ExportAllFrom
Implementation of interface ExportAllFrom.

ExportDefault
Implementation of interface ExportDefault.

ExportFrom
Implementation of interface ExportFrom.

ExportFromSpecifier
Implementation of interface ExportFromSpecifier.

ExportLocalSpecifier
Implementation of interface ExportLocalSpecifier.

ExportLocals
Implementation of interface ExportLocals.

ExpressionStatement
Implementation of interface ExpressionStatement.

ForInOfBinding
Implementation of interface ForInOfBinding.

ForInStatement
Implementation of interface ForInStatement.

ForOfStatement
Implementation of interface ForOfStatement.

ForStatement
Implementation of interface ForStatement.

FormalParameters
Implementation of interface FormalParameters.

FunctionExpressionContents
Implementation of interface FunctionExpressionContents.

FunctionOrMethodContents
Implementation of interface FunctionOrMethodContents.

GetterContents
Implementation of interface GetterContents.

IdentifierExpression
Implementation of interface IdentifierExpression.

IfStatement
Implementation of interface IfStatement.

Import
Implementation of interface Import.

ImportNamespace
Implementation of interface ImportNamespace.

ImportSpecifier
Implementation of interface ImportSpecifier.

LabelledStatement
Implementation of interface LabelledStatement.

LazyArrowExpressionWithExpression
Implementation of interface LazyArrowExpressionWithExpression.

LazyArrowExpressionWithFunctionBody
Implementation of interface LazyArrowExpressionWithFunctionBody.

LazyFunctionDeclaration
Implementation of interface LazyFunctionDeclaration.

LazyFunctionExpression
Implementation of interface LazyFunctionExpression.

LazyGetter
Implementation of interface LazyGetter.

LazyMethod
Implementation of interface LazyMethod.

LazySetter
Implementation of interface LazySetter.

LiteralBooleanExpression
Implementation of interface LiteralBooleanExpression.

LiteralInfinityExpression
Implementation of interface LiteralInfinityExpression.

LiteralNullExpression
Implementation of interface LiteralNullExpression.

LiteralNumericExpression
Implementation of interface LiteralNumericExpression.

LiteralPropertyName
Implementation of interface LiteralPropertyName.

LiteralRegExpExpression
Implementation of interface LiteralRegExpExpression.

LiteralStringExpression
Implementation of interface LiteralStringExpression.

Module
Implementation of interface Module.

NewExpression
Implementation of interface NewExpression.

NewTargetExpression
Implementation of interface NewTargetExpression.

Null
Implementation of interface .

ObjectAssignmentTarget
Implementation of interface ObjectAssignmentTarget.

ObjectBinding
Implementation of interface ObjectBinding.

ObjectExpression
Implementation of interface ObjectExpression.

ReturnStatement
Implementation of interface ReturnStatement.

Script
Implementation of interface Script.

SetterContents
Implementation of interface SetterContents.

ShorthandProperty
Implementation of interface ShorthandProperty.

SpreadElement
Implementation of interface SpreadElement.

StaticMemberAssignmentTarget
Implementation of interface StaticMemberAssignmentTarget.

StaticMemberExpression
Implementation of interface StaticMemberExpression.

Super
Implementation of interface Super.

SwitchCase
Implementation of interface SwitchCase.

SwitchDefault
Implementation of interface SwitchDefault.

SwitchStatement
Implementation of interface SwitchStatement.

SwitchStatementWithDefault
Implementation of interface SwitchStatementWithDefault.

TemplateElement
Implementation of interface TemplateElement.

TemplateExpression
Implementation of interface TemplateExpression.

ThisExpression
Implementation of interface ThisExpression.

ThrowStatement
Implementation of interface ThrowStatement.

TryCatchStatement
Implementation of interface TryCatchStatement.

TryFinallyStatement
Implementation of interface TryFinallyStatement.

UnaryExpression
Implementation of interface UnaryExpression.

UpdateExpression
Implementation of interface UpdateExpression.

VariableDeclaration
Implementation of interface VariableDeclaration.

VariableDeclarator
Implementation of interface VariableDeclarator.

ViewMutArguments
Shallow casting mechanism.

ViewMutArrayAssignmentTarget
Shallow casting mechanism.

ViewMutArrayBinding
Shallow casting mechanism.

ViewMutArrayExpression
Shallow casting mechanism.

ViewMutArrowExpressionContentsWithExpression
Shallow casting mechanism.

ViewMutArrowExpressionContentsWithFunctionBody
Shallow casting mechanism.

ViewMutAssertedBlockScope
Shallow casting mechanism.

ViewMutAssertedBoundName
Shallow casting mechanism.

ViewMutAssertedBoundNamesScope
Shallow casting mechanism.

ViewMutAssertedDeclaredName
Shallow casting mechanism.

ViewMutAssertedParameterName
Shallow casting mechanism.

ViewMutAssertedParameterScope
Shallow casting mechanism.

ViewMutAssertedPositionalParameterName
Shallow casting mechanism.

ViewMutAssertedRestParameterName
Shallow casting mechanism.

ViewMutAssertedScriptGlobalScope
Shallow casting mechanism.

ViewMutAssertedVarScope
Shallow casting mechanism.

ViewMutAssignmentExpression
Shallow casting mechanism.

ViewMutAssignmentTargetIdentifier
Shallow casting mechanism.

ViewMutAssignmentTargetPropertyIdentifier
Shallow casting mechanism.

ViewMutAssignmentTargetPropertyProperty
Shallow casting mechanism.

ViewMutAssignmentTargetWithInitializer
Shallow casting mechanism.

ViewMutAwaitExpression
Shallow casting mechanism.

ViewMutBinaryExpression
Shallow casting mechanism.

ViewMutBindingIdentifier
Shallow casting mechanism.

ViewMutBindingPropertyIdentifier
Shallow casting mechanism.

ViewMutBindingPropertyProperty
Shallow casting mechanism.

ViewMutBindingWithInitializer
Shallow casting mechanism.

ViewMutBlock
Shallow casting mechanism.

ViewMutBreakStatement
Shallow casting mechanism.

ViewMutCallExpression
Shallow casting mechanism.

ViewMutCatchClause
Shallow casting mechanism.

ViewMutClassDeclaration
Shallow casting mechanism.

ViewMutClassElement
Shallow casting mechanism.

ViewMutClassExpression
Shallow casting mechanism.

ViewMutCompoundAssignmentExpression
Shallow casting mechanism.

ViewMutComputedMemberAssignmentTarget
Shallow casting mechanism.

ViewMutComputedMemberExpression
Shallow casting mechanism.

ViewMutComputedPropertyName
Shallow casting mechanism.

ViewMutConditionalExpression
Shallow casting mechanism.

ViewMutContinueStatement
Shallow casting mechanism.

ViewMutDataProperty
Shallow casting mechanism.

ViewMutDebuggerStatement
Shallow casting mechanism.

ViewMutDirective
Shallow casting mechanism.

ViewMutDoWhileStatement
Shallow casting mechanism.

ViewMutEagerArrowExpressionWithExpression
Shallow casting mechanism.

ViewMutEagerArrowExpressionWithFunctionBody
Shallow casting mechanism.

ViewMutEagerFunctionDeclaration
Shallow casting mechanism.

ViewMutEagerFunctionExpression
Shallow casting mechanism.

ViewMutEagerGetter
Shallow casting mechanism.

ViewMutEagerMethod
Shallow casting mechanism.

ViewMutEagerSetter
Shallow casting mechanism.

ViewMutEmptyStatement
Shallow casting mechanism.

ViewMutExport
Shallow casting mechanism.

ViewMutExportAllFrom
Shallow casting mechanism.

ViewMutExportDefault
Shallow casting mechanism.

ViewMutExportFrom
Shallow casting mechanism.

ViewMutExportFromSpecifier
Shallow casting mechanism.

ViewMutExportLocalSpecifier
Shallow casting mechanism.

ViewMutExportLocals
Shallow casting mechanism.

ViewMutExpressionStatement
Shallow casting mechanism.

ViewMutForInOfBinding
Shallow casting mechanism.

ViewMutForInStatement
Shallow casting mechanism.

ViewMutForOfStatement
Shallow casting mechanism.

ViewMutForStatement
Shallow casting mechanism.

ViewMutFormalParameters
Shallow casting mechanism.

ViewMutFunctionExpressionContents
Shallow casting mechanism.

ViewMutFunctionOrMethodContents
Shallow casting mechanism.

ViewMutGetterContents
Shallow casting mechanism.

ViewMutIdentifierExpression
Shallow casting mechanism.

ViewMutIfStatement
Shallow casting mechanism.

ViewMutImport
Shallow casting mechanism.

ViewMutImportNamespace
Shallow casting mechanism.

ViewMutImportSpecifier
Shallow casting mechanism.

ViewMutLabel
Shallow casting mechanism for Label.

ViewMutLabelledStatement
Shallow casting mechanism.

ViewMutLazyArrowExpressionWithExpression
Shallow casting mechanism.

ViewMutLazyArrowExpressionWithFunctionBody
Shallow casting mechanism.

ViewMutLazyFunctionDeclaration
Shallow casting mechanism.

ViewMutLazyFunctionExpression
Shallow casting mechanism.

ViewMutLazyGetter
Shallow casting mechanism.

ViewMutLazyMethod
Shallow casting mechanism.

ViewMutLazySetter
Shallow casting mechanism.

ViewMutListOfAssertedBoundName
Shallow casting mechanism.

ViewMutListOfAssertedDeclaredName
Shallow casting mechanism.

ViewMutListOfAssertedMaybePositionalParameterName
Shallow casting mechanism.

ViewMutListOfAssignmentTargetOrAssignmentTargetWithInitializer
Shallow casting mechanism.

ViewMutListOfAssignmentTargetProperty
Shallow casting mechanism.

ViewMutListOfBindingProperty
Shallow casting mechanism.

ViewMutListOfClassElement
Shallow casting mechanism.

ViewMutListOfDirective
Shallow casting mechanism.

ViewMutListOfExportFromSpecifier
Shallow casting mechanism.

ViewMutListOfExportLocalSpecifier
Shallow casting mechanism.

ViewMutListOfExpressionOrTemplateElement
Shallow casting mechanism.

ViewMutListOfImportDeclarationOrExportDeclarationOrStatement
Shallow casting mechanism.

ViewMutListOfImportSpecifier
Shallow casting mechanism.

ViewMutListOfObjectProperty
Shallow casting mechanism.

ViewMutListOfOptionalBindingOrBindingWithInitializer
Shallow casting mechanism.

ViewMutListOfOptionalSpreadElementOrExpression
Shallow casting mechanism.

ViewMutListOfParameter
Shallow casting mechanism.

ViewMutListOfStatement
Shallow casting mechanism.

ViewMutListOfSwitchCase
Shallow casting mechanism.

ViewMutListOfVariableDeclarator
Shallow casting mechanism.

ViewMutLiteralBooleanExpression
Shallow casting mechanism.

ViewMutLiteralInfinityExpression
Shallow casting mechanism.

ViewMutLiteralNullExpression
Shallow casting mechanism.

ViewMutLiteralNumericExpression
Shallow casting mechanism.

ViewMutLiteralPropertyName
Shallow casting mechanism.

ViewMutLiteralRegExpExpression
Shallow casting mechanism.

ViewMutLiteralStringExpression
Shallow casting mechanism.

ViewMutModule
Shallow casting mechanism.

ViewMutNewExpression
Shallow casting mechanism.

ViewMutNewTargetExpression
Shallow casting mechanism.

ViewMutNull
Shallow casting mechanism.

ViewMutObjectAssignmentTarget
Shallow casting mechanism.

ViewMutObjectBinding
Shallow casting mechanism.

ViewMutObjectExpression
Shallow casting mechanism.

ViewMutOffset

ViewMutOptionalAssignmentTarget
Shallow casting mechanism.

ViewMutOptionalBinding
Shallow casting mechanism.

ViewMutOptionalBindingIdentifier
Shallow casting mechanism.

ViewMutOptionalBindingOrBindingWithInitializer
Shallow casting mechanism.

ViewMutOptionalCatchClause
Shallow casting mechanism.

ViewMutOptionalExpression
Shallow casting mechanism.

ViewMutOptionalIdentifierName
Shallow casting mechanism.

ViewMutOptionalLabel
Shallow casting mechanism.

ViewMutOptionalPropertyKey
Shallow casting mechanism.

ViewMutOptionalSpreadElementOrExpression
Shallow casting mechanism.

ViewMutOptionalStatement
Shallow casting mechanism.

ViewMutOptionalVariableDeclarationOrExpression
Shallow casting mechanism.

ViewMutReturnStatement
Shallow casting mechanism.

ViewMutScript
Shallow casting mechanism.

ViewMutSetterContents
Shallow casting mechanism.

ViewMutShorthandProperty
Shallow casting mechanism.

ViewMutSpreadElement
Shallow casting mechanism.

ViewMutStaticMemberAssignmentTarget
Shallow casting mechanism.

ViewMutStaticMemberExpression
Shallow casting mechanism.

ViewMutString
Shallow casting mechanism for String.

ViewMutSuper
Shallow casting mechanism.

ViewMutSwitchCase
Shallow casting mechanism.

ViewMutSwitchDefault
Shallow casting mechanism.

ViewMutSwitchStatement
Shallow casting mechanism.

ViewMutSwitchStatementWithDefault
Shallow casting mechanism.

ViewMutTemplateElement
Shallow casting mechanism.

ViewMutTemplateExpression
Shallow casting mechanism.

ViewMutThisExpression
Shallow casting mechanism.

ViewMutThrowStatement
Shallow casting mechanism.

ViewMutTryCatchStatement
Shallow casting mechanism.

ViewMutTryFinallyStatement
Shallow casting mechanism.

ViewMutUnaryExpression
Shallow casting mechanism.

ViewMutUpdateExpression
Shallow casting mechanism.

ViewMutVariableDeclaration
Shallow casting mechanism.

ViewMutVariableDeclarator
Shallow casting mechanism.

ViewMutWhileStatement
Shallow casting mechanism.

ViewMutWithStatement
Shallow casting mechanism.

ViewMutYieldExpression
Shallow casting mechanism.

ViewMutYieldStarExpression
Shallow casting mechanism.

WhileStatement
Implementation of interface WhileStatement.

WithStatement
Implementation of interface WithStatement.

YieldExpression
Implementation of interface YieldExpression.

YieldStarExpression
Implementation of interface YieldStarExpression.

Enums

ASTField
All the fields.

ASTNode
All the interfaces.

ArrowExpression
Implementation of interface sum ArrowExpression

AssertedDeclaredKind
Implementation of string enum AssertedDeclaredKind

AssertedMaybePositionalParameterName
Implementation of interface sum AssertedMaybePositionalParameterName

AssignmentTarget
Implementation of interface sum AssignmentTarget

AssignmentTargetOrAssignmentTargetWithInitializer
Implementation of interface sum AssignmentTargetOrAssignmentTargetWithInitializer

AssignmentTargetPattern
Implementation of interface sum AssignmentTargetPattern

AssignmentTargetProperty
Implementation of interface sum AssignmentTargetProperty

BinaryOperator
Implementation of string enum BinaryOperator

Binding
Implementation of interface sum Binding

BindingOrBindingWithInitializer
Implementation of interface sum BindingOrBindingWithInitializer

BindingPattern
Implementation of interface sum BindingPattern

BindingProperty
Implementation of interface sum BindingProperty

CompoundAssignmentOperator
Implementation of string enum CompoundAssignmentOperator

ExportDeclaration
Implementation of interface sum ExportDeclaration

Expression
Implementation of interface sum Expression

ExpressionOrSuper
Implementation of interface sum ExpressionOrSuper

ExpressionOrTemplateElement
Implementation of interface sum ExpressionOrTemplateElement

ForInOfBindingOrAssignmentTarget
Implementation of interface sum ForInOfBindingOrAssignmentTarget

FunctionDeclaration
Implementation of interface sum FunctionDeclaration

FunctionDeclarationOrClassDeclarationOrExpression
Implementation of interface sum FunctionDeclarationOrClassDeclarationOrExpression

FunctionDeclarationOrClassDeclarationOrVariableDeclaration
Implementation of interface sum FunctionDeclarationOrClassDeclarationOrVariableDeclaration

FunctionExpression
Implementation of interface sum FunctionExpression

Getter
Implementation of interface sum Getter

ImportDeclaration
Implementation of interface sum ImportDeclaration

ImportDeclarationOrExportDeclarationOrStatement
Implementation of interface sum ImportDeclarationOrExportDeclarationOrStatement

IterationStatement
Implementation of interface sum IterationStatement

Literal
Implementation of interface sum Literal

Method
Implementation of interface sum Method

MethodDefinition
Implementation of interface sum MethodDefinition

ObjectProperty
Implementation of interface sum ObjectProperty

Parameter
Implementation of interface sum Parameter

Program
Implementation of interface sum Program

PropertyName
Implementation of interface sum PropertyName

Setter
Implementation of interface sum Setter

SimpleAssignmentTarget
Implementation of interface sum SimpleAssignmentTarget

SpreadElementOrExpression
Implementation of interface sum SpreadElementOrExpression

Statement
Implementation of interface sum Statement

UnaryOperator
Implementation of string enum UnaryOperator

UpdateOperator
Implementation of string enum UpdateOperator

VariableDeclarationKind
Implementation of string enum VariableDeclarationKind

VariableDeclarationOrExpression
Implementation of interface sum VariableDeclarationOrExpression

ViewMutArrowExpression
A mechanism to view value as an instance of interface sum ArrowExpression

ViewMutAssertedMaybePositionalParameterName
A mechanism to view value as an instance of interface sum AssertedMaybePositionalParameterName

ViewMutAssignmentTarget
A mechanism to view value as an instance of interface sum AssignmentTarget

ViewMutAssignmentTargetOrAssignmentTargetWithInitializer
A mechanism to view value as an instance of interface sum AssignmentTargetOrAssignmentTargetWithInitializer

ViewMutAssignmentTargetPattern
A mechanism to view value as an instance of interface sum AssignmentTargetPattern

ViewMutAssignmentTargetProperty
A mechanism to view value as an instance of interface sum AssignmentTargetProperty

ViewMutBinding
A mechanism to view value as an instance of interface sum Binding

ViewMutBindingOrBindingWithInitializer
A mechanism to view value as an instance of interface sum BindingOrBindingWithInitializer

ViewMutBindingPattern
A mechanism to view value as an instance of interface sum BindingPattern

ViewMutBindingProperty
A mechanism to view value as an instance of interface sum BindingProperty

ViewMutExportDeclaration
A mechanism to view value as an instance of interface sum ExportDeclaration

ViewMutExpression
A mechanism to view value as an instance of interface sum Expression

ViewMutExpressionOrSuper
A mechanism to view value as an instance of interface sum ExpressionOrSuper

ViewMutExpressionOrTemplateElement
A mechanism to view value as an instance of interface sum ExpressionOrTemplateElement

ViewMutForInOfBindingOrAssignmentTarget
A mechanism to view value as an instance of interface sum ForInOfBindingOrAssignmentTarget

ViewMutFunctionDeclaration
A mechanism to view value as an instance of interface sum FunctionDeclaration

ViewMutFunctionDeclarationOrClassDeclarationOrExpression
A mechanism to view value as an instance of interface sum FunctionDeclarationOrClassDeclarationOrExpression

ViewMutFunctionDeclarationOrClassDeclarationOrVariableDeclaration
A mechanism to view value as an instance of interface sum FunctionDeclarationOrClassDeclarationOrVariableDeclaration

ViewMutFunctionExpression
A mechanism to view value as an instance of interface sum FunctionExpression

ViewMutGetter
A mechanism to view value as an instance of interface sum Getter

ViewMutImportDeclaration
A mechanism to view value as an instance of interface sum ImportDeclaration

ViewMutImportDeclarationOrExportDeclarationOrStatement
A mechanism to view value as an instance of interface sum ImportDeclarationOrExportDeclarationOrStatement

ViewMutIterationStatement
A mechanism to view value as an instance of interface sum IterationStatement

ViewMutLiteral
A mechanism to view value as an instance of interface sum Literal

ViewMutMethod
A mechanism to view value as an instance of interface sum Method

ViewMutMethodDefinition
A mechanism to view value as an instance of interface sum MethodDefinition

ViewMutObjectProperty
A mechanism to view value as an instance of interface sum ObjectProperty

ViewMutParameter
A mechanism to view value as an instance of interface sum Parameter

ViewMutProgram
A mechanism to view value as an instance of interface sum Program

ViewMutPropertyName
A mechanism to view value as an instance of interface sum PropertyName

ViewMutSetter
A mechanism to view value as an instance of interface sum Setter

ViewMutSimpleAssignmentTarget
A mechanism to view value as an instance of interface sum SimpleAssignmentTarget

ViewMutSpreadElementOrExpression
A mechanism to view value as an instance of interface sum SpreadElementOrExpression

ViewMutStatement
A mechanism to view value as an instance of interface sum Statement

ViewMutVariableDeclarationOrExpression
A mechanism to view value as an instance of interface sum VariableDeclarationOrExpression

Traits

Visitor
A set of callbacks used to inspect the contents of an AST in a strongly-typed manner. For each node Foo, enter_foo() will be called before visiting the children, giving the opportunity to alter the node, and enter_foo() will be called after visiting the children, giving the opportunity to alter it further.

Walker

Type Definitions

Arguments
Implementation of list type Arguments.

FunctionBody

IOPath
A Path, used when walking the tree with more weakly-typed APIs, e.g. TokenReader/TokenWriter.

Label
Alias to primitive type.

ListOfAssertedBoundName
Implementation of list type ListOfAssertedBoundName.

ListOfAssertedDeclaredName
Implementation of list type ListOfAssertedDeclaredName.

ListOfAssertedMaybePositionalParameterName
Implementation of list type ListOfAssertedMaybePositionalParameterName.

ListOfAssignmentTargetOrAssignmentTargetWithInitializer
Implementation of list type ListOfAssignmentTargetOrAssignmentTargetWithInitializer.

ListOfAssignmentTargetProperty
Implementation of list type ListOfAssignmentTargetProperty.

ListOfBindingProperty
Implementation of list type ListOfBindingProperty.

ListOfClassElement
Implementation of list type ListOfClassElement.

ListOfDirective
Implementation of list type ListOfDirective.

ListOfExportFromSpecifier
Implementation of list type ListOfExportFromSpecifier.

ListOfExportLocalSpecifier
Implementation of list type ListOfExportLocalSpecifier.

ListOfExpressionOrTemplateElement
Implementation of list type ListOfExpressionOrTemplateElement.

ListOfImportDeclarationOrExportDeclarationOrStatement
Implementation of list type ListOfImportDeclarationOrExportDeclarationOrStatement.

ListOfImportSpecifier
Implementation of list type ListOfImportSpecifier.

ListOfObjectProperty
Implementation of list type ListOfObjectProperty.

ListOfOptionalBindingOrBindingWithInitializer
Implementation of list type ListOfOptionalBindingOrBindingWithInitializer.

ListOfOptionalSpreadElementOrExpression
Implementation of list type ListOfOptionalSpreadElementOrExpression.

ListOfParameter
Implementation of list type ListOfParameter.

ListOfStatement
Implementation of list type ListOfStatement.

ListOfSwitchCase
Implementation of list type ListOfSwitchCase.

ListOfVariableDeclarator
Implementation of list type ListOfVariableDeclarator.

OptionalAssignmentTarget
Alias to optional type type.

OptionalBinding
Alias to optional type type.

OptionalBindingIdentifier
Alias to optional type type.

OptionalBindingOrBindingWithInitializer
Alias to optional type type.

OptionalCatchClause
Alias to optional type type.

OptionalExpression
Alias to optional type type.

OptionalIdentifierName
Alias to optional type type.

OptionalLabel
Alias to optional type type.

OptionalPropertyKey
Alias to optional type type.

OptionalSpreadElementOrExpression
Alias to optional type type.

OptionalStatement
Alias to optional type type.

OptionalVariableDeclarationOrExpression
Alias to optional type type.

String
Alias to primitive type.

ViewMutFunctionBody

WalkPath
A Path, used when walking the tree with the strongly-typed Walker API.

WalkPathItem
A PathItem, used when walking the tree with the strongly-typed Walker API.