Enum binjs_es6::ast::ViewMutExpressionOrTemplateElement

pub enum ViewMutExpressionOrTemplateElement<'a> {
    CallExpression(&'a mut CallExpression),
    EagerArrowExpressionWithExpression(&'a mut EagerArrowExpressionWithExpression),
    LiteralInfinityExpression(&'a mut LiteralInfinityExpression),
    LazyArrowExpressionWithFunctionBody(&'a mut LazyArrowExpressionWithFunctionBody),
    AssignmentExpression(&'a mut AssignmentExpression),
    TemplateElement(&'a mut TemplateElement),
    NewExpression(&'a mut NewExpression),
    ArrayExpression(&'a mut ArrayExpression),
    LazyFunctionExpression(&'a mut LazyFunctionExpression),
    ComputedMemberExpression(&'a mut ComputedMemberExpression),
    LazyArrowExpressionWithExpression(&'a mut LazyArrowExpressionWithExpression),
    EagerArrowExpressionWithFunctionBody(&'a mut EagerArrowExpressionWithFunctionBody),
    CompoundAssignmentExpression(&'a mut CompoundAssignmentExpression),
    ThisExpression(&'a mut ThisExpression),
    UpdateExpression(&'a mut UpdateExpression),
    StaticMemberExpression(&'a mut StaticMemberExpression),
    LiteralBooleanExpression(&'a mut LiteralBooleanExpression),
    ObjectExpression(&'a mut ObjectExpression),
    IdentifierExpression(&'a mut IdentifierExpression),
    LiteralNullExpression(&'a mut LiteralNullExpression),
    TemplateExpression(&'a mut TemplateExpression),
    YieldExpression(&'a mut YieldExpression),
    LiteralRegExpExpression(&'a mut LiteralRegExpExpression),
    UnaryExpression(&'a mut UnaryExpression),
    LiteralStringExpression(&'a mut LiteralStringExpression),
    YieldStarExpression(&'a mut YieldStarExpression),
    LiteralNumericExpression(&'a mut LiteralNumericExpression),
    NewTargetExpression(&'a mut NewTargetExpression),
    ClassExpression(&'a mut ClassExpression),
    AwaitExpression(&'a mut AwaitExpression),
    BinaryExpression(&'a mut BinaryExpression),
    EagerFunctionExpression(&'a mut EagerFunctionExpression),
    ConditionalExpression(&'a mut ConditionalExpression),
    BinASTExpressionWithProbabilityTable(&'a mut BinASTExpressionWithProbabilityTable),
}

A mechanism to view value as an instance of interface sum ExpressionOrTemplateElement

Used to perform shallow cast between larger sums and smaller sums.

Variants
CallExpression(&'a mut CallExpression)EagerArrowExpressionWithExpression(&'a mut EagerArrowExpressionWithExpression)LiteralInfinityExpression(&'a mut LiteralInfinityExpression)LazyArrowExpressionWithFunctionBody(&'a mut LazyArrowExpressionWithFunctionBody)AssignmentExpression(&'a mut AssignmentExpression)TemplateElement(&'a mut TemplateElement)NewExpression(&'a mut NewExpression)ArrayExpression(&'a mut ArrayExpression)LazyFunctionExpression(&'a mut LazyFunctionExpression)ComputedMemberExpression(&'a mut ComputedMemberExpression)LazyArrowExpressionWithExpression(&'a mut LazyArrowExpressionWithExpression)EagerArrowExpressionWithFunctionBody(&'a mut EagerArrowExpressionWithFunctionBody)CompoundAssignmentExpression(&'a mut CompoundAssignmentExpression)ThisExpression(&'a mut ThisExpression)UpdateExpression(&'a mut UpdateExpression)StaticMemberExpression(&'a mut StaticMemberExpression)LiteralBooleanExpression(&'a mut LiteralBooleanExpression)ObjectExpression(&'a mut ObjectExpression)IdentifierExpression(&'a mut IdentifierExpression)LiteralNullExpression(&'a mut LiteralNullExpression)TemplateExpression(&'a mut TemplateExpression)YieldExpression(&'a mut YieldExpression)LiteralRegExpExpression(&'a mut LiteralRegExpExpression)UnaryExpression(&'a mut UnaryExpression)LiteralStringExpression(&'a mut LiteralStringExpression)YieldStarExpression(&'a mut YieldStarExpression)LiteralNumericExpression(&'a mut LiteralNumericExpression)NewTargetExpression(&'a mut NewTargetExpression)ClassExpression(&'a mut ClassExpression)AwaitExpression(&'a mut AwaitExpression)BinaryExpression(&'a mut BinaryExpression)EagerFunctionExpression(&'a mut EagerFunctionExpression)ConditionalExpression(&'a mut ConditionalExpression)BinASTExpressionWithProbabilityTable(&'a mut BinASTExpressionWithProbabilityTable)
Methods

impl<'a> ViewMutExpressionOrTemplateElement<'a>

pub fn steal(&mut self) -> ExpressionOrTemplateElement

Convert a ViewMutExpressionOrTemplateElement back into a ExpressionOrTemplateElement, stealing the contents of the sum.